Added three new TrimMs functions to the Chebyshev Subdivision Solver
