--- yroots/TestArrays.py
@@ -1,4415 +1,0 @@
-import ChebyshevSubdivisionSolver as CS
-import numpy as np
-
-def test_getSubdivisionIntervals():
-    trackedInterval_1 = CS.TrackedInterval(np.array([[-0.63673497,  0.75388677],
-                                                    [-1.,          0.24771388],
-                                                    [-1.,          1.        ]]))
-    Ms_1 = [np.array([[[-1.19628962e-01,  4.78500000e+00,  1.36125000e+00],       
-                        [-1.22216380e+00, -5.88038994e-17, -1.29732624e-16]],       
-
-                    [[ 2.75448343e+01, -4.09625873e-15, -5.24707922e-16],        
-                        [ 1.38346959e+01, -1.64018392e-15, -9.14780575e-16]],       
-
-                    [[-2.34899407e+00, -5.34145567e-16, -3.02474728e-16],        
-                        [-1.00260753e-15,  1.73733322e-16,  3.97924039e-17]],       
-
-                    [[ 4.43170528e+00,  1.12302734e-16,  4.10866620e-16],        
-                        [ 9.31305763e-16, -1.86261153e-16, -1.86261153e-16]]]), 
-
-            np.array([[[ 2.84845613e+02,  1.65000000e+00],
-                        [ 2.27137940e+02, -1.72605562e-13],
-                        [ 1.29793108e+02,  2.54414728e-14],
-                        [ 4.86724156e+01,  6.09220633e-15],
-                        [ 1.08160924e+01,  1.01171420e-14],
-                        [ 1.08160924e+00, -3.13565083e-15]],
-
-                    [[-9.64091464e+01, -2.86049313e-14],
-                        [ 4.14691757e-14,  7.88666604e-14],
-                        [ 2.67503292e-14, -9.27142549e-15],
-                        [-6.34389840e-14, -6.91944204e-15],
-                        [-1.24333103e-14, -2.02584903e-15],
-                        [ 6.92587964e-15,  8.01648088e-16]],
-
-                    [[ 1.93541862e+02, -5.17408223e-14],
-                        [-5.39535253e-14, -1.96574197e-14],
-                        [-5.00488782e-14, -4.71474082e-14],
-                        [ 2.06908008e-14,  8.68344641e-15],
-                        [ 1.40966336e-14,  2.75871201e-14],
-                        [-3.57269884e-15, -6.35581641e-15]],
-
-                    [[-3.26656222e+01,  1.37816904e-14],
-                        [-1.57824001e-14, -1.92527602e-14],
-                        [ 1.07754349e-15,  7.17861036e-15],
-                        [ 2.10137810e-14, -6.69939041e-16],
-                        [-4.24524051e-15, -2.25717150e-16],
-                        [-3.59823275e-16,  3.49127175e-17]],
-
-                    [[ 4.62211928e+01,  4.21235493e-15],
-                        [-7.00784374e-16,  8.21968518e-15],
-                        [ 4.49894444e-15,  1.37359712e-14],
-                        [ 1.55416701e-15, -5.19175671e-15],
-                        [-2.74045293e-15, -6.88588808e-15],
-                        [ 4.70817493e-16,  1.72633081e-15]]]), 
-
-            np.array([[[ 4.68071678e+00, -4.78500000e+00, -1.36125000e+00],
-                        [ 6.25613546e+00, -2.64267077e-17, -1.19440696e-16],        
-                        [ 1.56403386e+00,  7.59312342e-18,  4.78427451e-17]],       
-
-                    [[-1.20118630e+00, -1.72239045e-16, -1.51455377e-16],        
-                        [ 5.52210332e-16, -3.13691772e-16, -1.22891472e-16],        
-                        [-8.91891458e-17,  9.01323150e-17,  3.09872741e-17]],       
-
-                    [[ 3.39930849e+00,  0.00000000e+00,  3.51928144e-17],        
-                        [ 1.00761893e-16,  0.00000000e+00,  1.00761893e-16],        
-                        [-4.17799856e-17,  0.00000000e+00, -4.17799856e-17]]])]
-    errors_1 = np.array([.000000132,.000000000050623,.00000000000203045])
-    level_1 = 1
-    expected_allMs_first_1 = [-10.63142908487176896415, 149.99524012496584646215, 3.58368898695362148032, -5.79272022070384640102, 151.64524012496596583333, -1.25501987721429886236, -18.49801308357789153547, 357.79251527102314867079, 9.90153425936189890422, -13.65930421940996808416, 359.44251527102306909001, 5.06282539519397811745, 7.97080004997253688259, 74.75050093156620789614, 2.51662426488145785086, 12.80950891414045500483, 76.40050093156625621305, -2.32208459928646249182, 13.93891195126641413538, 282.54777607762349589393, 8.83446953728973483067, 18.77762081543433225761, 284.19777607762341631314, 3.99576067312181493207]
-    expected_allMs_shapes_1 = [(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3),(4, 2, 3),(5, 6, 2),(3, 3, 3)]
-    expected_allErrors_1 = [np.array([0.00000013200011032514,0.00000000005289171023,0.00000000000206817728]),np.array([0.00000013200010927577,0.00000000005289171023,0.00000000000206625355]),np.array([0.00000013200011032514,0.00000000005371982276,0.00000000000207426074]),np.array([0.00000013200010927577,0.00000000005371982276,0.00000000000207111266]),np.array([0.00000013200009976412,0.00000000005277706881,0.00000000000206817728]),np.array([0.00000013200010301239,0.00000000005277706881,0.00000000000206625355]),np.array([0.00000013200009976412,0.00000000005360518133,0.00000000000207426074]),np.array([0.00000013200010301239,0.00000000005360518133,0.00000000000207111266])]    
-    expected_allIntervals_1 = [np.array([[-0.63673497,  0.08600977], [-1.,         -0.35152844], [-1.,          0.03945555]]),np.array([[-0.63673497,  0.08600977], [-1.,         -0.35152844], [ 0.03945555,  1.        ]]),np.array([[-0.63673497,  0.08600977], [-0.35152844,  0.24771388], [-1.,          0.03945555]]),np.array([[-0.63673497,  0.08600977], [-0.35152844,  0.24771388], [ 0.03945555,  1.        ]]),np.array([[ 0.08600977,  0.75388677], [-1.,         -0.35152844], [-1.,          0.03945555]]),np.array([[ 0.08600977,  0.75388677], [-1.,         -0.35152844], [ 0.03945555,  1.        ]]),np.array([[ 0.08600977,  0.75388677], [-0.35152844,  0.24771388], [-1.,  0.03945555]]),np.array([[ 0.08600977,  0.75388677], [-0.35152844,  0.24771388], [ 0.03945555,  1.        ]])]
-    expected_transforms_1 = [[np.array([[ 0.51972777,  1.,  1.        ],[-0.48027223,  0.,  0.        ]]), np.array([[ 1.,  0.51972777,  1.        ],[ 0., -0.48027223,  0.        ]]), np.array([[ 1.,  1.,  0.51972777],[ 0.,  0., -0.48027223]])],[np.array([[ 0.51972777,  1.,  1.        ],[-0.48027223,  0.,  0.        ]]), np.array([[ 1.,  0.51972777,  1.        ],[ 0., -0.48027223,  0.        ]]), np.array([[1., 1., 0.48027223],[0., 0., 0.51972777]])],[np.array([[ 0.51972777,  1.,  1.        ],[-0.48027223,  0.,  0.        ]]), np.array([[1., 0.48027223, 1.        ],[0., 0.51972777, 0.        ]]), np.array([[ 1.,  1.,  0.51972777],[ 0.,  0., -0.48027223]])],[np.array([[ 0.51972777,  1.,  1.        ],[-0.48027223,  0.,  0.        ]]), np.array([[1., 0.48027223, 1.        ],[0., 0.51972777, 0.        ]]), np.array([[1., 1., 0.48027223],[0., 0., 0.51972777]])],[np.array([[0.48027223, 1., 1.        ],[0.51972777, 0., 0.        ]]), np.array([[ 1.,  0.51972777,  1.        ],[ 0., -0.48027223,  0.        ]]), np.array([[ 1.,  1.,  0.51972777],[ 0.,  0., -0.48027223]])],[np.array([[0.48027223, 1., 1.        ],[0.51972777, 0., 0.        ]]), np.array([[ 1.,  0.51972777,  1.        ],[ 0., -0.48027223,  0.        ]]), np.array([[1., 1., 0.48027223],[0., 0., 0.51972777]])],[np.array([[0.48027223, 1., 1.        ],[0.51972777, 0., 0.        ]]), np.array([[1., 0.48027223, 1.        ],[0., 0.51972777, 0.        ]]), np.array([[ 1.,  1.,  0.51972777],[ 0.,  0., -0.48027223]])],[np.array([[0.48027223, 1., 1.        ],[0.51972777, 0., 0.        ]]), np.array([[1., 0.48027223, 1.        ],[0., 0.51972777, 0.        ]]), np.array([[1., 1., 0.48027223],[0., 0., 0.51972777]])]]
-    allMs_1, allErrors_1, allIntervals_1 = CS.getSubdivisionIntervals(Ms_1,errors_1,trackedInterval_1,False,level_1)
-    allMs_first_1 = [M[0,0,0] for Ms in allMs_1 for M in Ms]
-    allMs_shapes_1 = [M.shape for Ms in allMs_1 for M in Ms]
-    allTransforms_1 = [interval.transforms for interval in allIntervals_1]
-    assert np.allclose(expected_allMs_first_1,allMs_first_1)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_1,expected_allMs_shapes_1)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_1,expected_allErrors_1)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_1,expected_allIntervals_1))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_1,expected_transforms_1))
-
-    trackedInterval_2 = CS.TrackedInterval(np.array([[ 0.03945555,  0.03945555 + 1e-5        ],
-                                                        [-1.,          -1 + 1e-5],
-                                                        [ 1.,  1. + 1e-5        ],
-                                                        [ 0.03945555,  0.03945555 + 1e-5 ],
-                                                        [ 0.41480102,  0.41480102 + 1e-5       ]]))
-    Ms_2 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-    errors_2 = np.array([.00000453450132,.0000000002050623,.00000000005540203045,.00000000005543045,.00000000000203045])
-    level_2 = 2
-    expected_allMs_first_2 = [0.8213663176493646, 0.9384350034288482, 0.8724926375293429, 0.23148354651427472, 0.39466226188438286, 0.7134088515909089, 1.1232730105663322, 1.057330644666827, 0.2912971276928736, 0.5795002690218667, 0.6091824841909089, 1.395269813752024, 1.1009100432106587, 0.1532064117422049, 0.6225642568843829, 1.0628440611324532, 1.580107820889508, 1.2857480503481424, 0.42020792410608393, 0.8074022640218667, 1.3263464365014803, 1.4434151222809637, 1.3774727563814584, 0.7364636653663904, 0.8996423807364986, 1.2183889704430244, 1.6282531294184477, 1.5623107635189424, 0.7962772465449893, 1.0844803878739822, 1.1141626030430245, 1.9002499326041395, 1.6058901620627741, 0.6581865305943204, 1.1275443757364985, 1.5678241799845687, 2.0850879397416238, 1.7907281692002581, 0.9251880429581996, 1.312382382873982]
-    expected_allMs_shapes_2 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_2 = [np.array([0.00000453450132485112, 0.00000000020507035934, 0.00000000005540911298, 0.00000000005543325050, 0.00000000000203521656]), np.array([0.00000453450132497124, 0.00000000020507080860, 0.00000000005540956224, 0.00000000005543325050, 0.00000000000203559094]), np.array([0.00000453450132485112, 0.00000000020507094781, 0.00000000005540940722, 0.00000000005543374089, 0.00000000000203521656]), np.array([0.00000453450132497124, 0.00000000020507139707, 0.00000000005540985648, 0.00000000005543374089, 0.00000000000203559094]), np.array([0.00000453450132550317, 0.00000000020507035934, 0.00000000005540911298, 0.00000000005543357653, 0.00000000000203554259]), np.array([0.00000453450132562329, 0.00000000020507080860, 0.00000000005540956224, 0.00000000005543357653, 0.00000000000203591697]), np.array([0.00000453450132550317, 0.00000000020507094781, 0.00000000005540940722, 0.00000000005543406692, 0.00000000000203554259]), np.array([0.00000453450132562329, 0.00000000020507139707, 0.00000000005540985648, 0.00000000005543406692, 0.00000000000203591697])]
-    expected_allIntervals_2 = [np.array([[ 0.03945555,  0.03946075], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03945555,  0.03946075], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03945555,  0.03946075], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03945555,  0.03946075], [ 0.41480622,  0.41481102]]),  np.array([[ 0.03945555,  0.03946075], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03946075,  0.03946555], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03945555,  0.03946075], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03946075,  0.03946555], [ 0.41480622,  0.41481102]]),  np.array([[ 0.03946075,  0.03946555], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03945555,  0.03946075], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03946075,  0.03946555], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03945555,  0.03946075], [ 0.41480622,  0.41481102]]),  np.array([[ 0.03946075,  0.03946555], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03946075,  0.03946555], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03946075,  0.03946555], [-1.,         -0.99999   ], [ 1.,          1.00001   ], [ 0.03946075,  0.03946555], [ 0.41480622,  0.41481102]])]
-    expected_transforms_2 = [[np.array([[ 0.51972777,  1.,  1.,  1.,  1.],[-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.],[ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777],[ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.],[-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.],[ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223],[0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.],[-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.],[0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777],[ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.],[-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.],[0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223],[0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.],[0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.],[ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777],[ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.],[0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.],[ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223],[0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.],[0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.],[0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777],[ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.],[0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.],[0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223],[0., 0., 0., 0., 0.51972777]])]]
-    allMs_2, allErrors_2, allIntervals_2 = CS.getSubdivisionIntervals(Ms_2,errors_2,trackedInterval_2,False,level_2)
-    allMs_first_2 = [M[0,0,0,0,0] for Ms in allMs_2 for M in Ms]
-    allMs_shapes_2 = [M.shape for Ms in allMs_2 for M in Ms]
-    allTransforms_2 = [interval.transforms for interval in allIntervals_2]
-    assert np.allclose(expected_allMs_first_2,allMs_first_2)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_2,expected_allMs_shapes_2)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_2,expected_allErrors_2)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_2,expected_allIntervals_2))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_2,expected_transforms_2))
-
-
-    trackedInterval_3 = CS.TrackedInterval(np.array([[ 0.03945555,  1.3828272        ],
-                                                        [-1.,          -1 + 1e-5],
-                                                        [ 1.,  1.532344       ],
-                                                        [ 0.03945555,  0.03945555 + 1e-5 ],
-                                                        [ 0.41480102,  0.41480102 + 1e-5       ]]))
-    Ms_3 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-    errors_3 = np.array([.00000000000123023030,.000000000030202220202,.0000012332,.0032123123,.0000000377321])
-    level_3 = 2
-    expected_allMs_first_3 = [0.56464181584606421005, 1.21247573584606427310, 0.98898473584606416686, 0.24695156289991765375, 0.65410566794005731150, 0.87452136050034234849, 1.52235528050034218950, 1.29886428050034230530, 0.28413710829448540762, 0.70937223818887207294, 1.06962193469817967717, 1.71745585469817974023, 1.49396485469817963398, 0.75193168175203328740, 1.15908578679217266760, 1.37950147935245781561, 2.02733539935245810071, 1.80384439935245777242, 0.78911722714660104128, 1.21435235704098776210]
-    expected_allMs_shapes_3 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_3 = [np.array([0.00000000000123465505, 0.00000000003020788970, 0.00000123320000482213, 0.00321231230000266522, 0.00000003773210378504]), np.array([0.00000000000123505026, 0.00000000003020828492, 0.00000123320000521735, 0.00321231230000271293, 0.00000003773210410023]), np.array([0.00000000000123465505, 0.00000000003020788970, 0.00000123320000482213, 0.00321231230000266522, 0.00000003773210378504]), np.array([0.00000000000123505026, 0.00000000003020828492, 0.00000123320000521735, 0.00321231230000271293, 0.00000003773210410023])]
-    expected_allIntervals_3 = [np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03945555,  0.03946555], [ 0.41480102,  0.41481102]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03945555,  0.03946555], [ 0.41480102,  0.41481102]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03945555,  0.03946555], [ 0.41480102,  0.41481102]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03945555,  0.03946555], [ 0.41480102,  0.41481102]])]
-    expected_transforms_3 = [[np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]])]]
-    allMs_3, allErrors_3, allIntervals_3 = CS.getSubdivisionIntervals(Ms_3,errors_3,trackedInterval_3,False,level_3)
-    allMs_first_3 = [M[0,0,0,0,0] for Ms in allMs_3 for M in Ms]
-    allMs_shapes_3 = [M.shape for Ms in allMs_3 for M in Ms]
-    allTransforms_3 = [interval.transforms for interval in allIntervals_3]
-    assert np.allclose(expected_allMs_first_3,allMs_first_3)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_3,expected_allMs_shapes_3)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_3,expected_allErrors_3)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_3,expected_allIntervals_3))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_3,expected_transforms_3))
-
-    trackedInterval_4 = CS.TrackedInterval(np.array([[ 0.03945555,  1.3828272        ],
-                                                        [-1.,          -1 + 1e-5],
-                                                        [ 1.,  1.532344       ],
-                                                        [ 0.03945555,  0.03945555 + 1e-5 ],
-                                                        [ 0.41480102,  0.41480102 + 1e-5       ]]))
-    Ms_4 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-    errors_4 = np.array([.00001323,.0073467,.00000000763456,.00000000456,.00000000234534])
-    level_4 = 6
-    expected_allMs_first_4 = [0.58885294941503707467, 0.70592163519452066911, 0.63997926929501536275, 0.20358194233380055849, 0.42233584209028984802, 0.48089548335658127698, 0.89075964233200455311, 0.82481727643249924675, 0.26339552351239942096, 0.60717384922777373202, 0.37666911595658125744, 1.16275644551769641311, 0.86839667497633099558, 0.12530480756173073553, 0.50627364225610382853, 0.83033069289812544422, 1.34759445265518040813, 1.05323468211381499060, 0.39230631992560976995, 0.69111164939358771253, 0.89873249406931510208, 1.01580117984879847448, 0.94985881394929361221, 0.24076748772836842338, 0.38545419297232186295, 0.79077502801085941542, 1.20063918698628269155, 1.13469682108677738519, 0.30058106890696728586, 0.57029220010980563593, 0.68654866061085939588, 1.47263599017197432950, 1.17827621963060913401, 0.16249035295629843390, 0.66125863530417194891, 1.14021023755240369368, 1.65747399730945854657, 1.36311422676809312904, 0.42949186532017746831, 0.84609664244165572189, 1.09383306826715243076, 1.21090175404663602521, 1.14495938814713071885, 0.70856206118591624765, 0.92731596094240553718, 0.98587560220869674410, 1.39573976118412002023, 1.32979739528461471387, 0.76837564236451516564, 1.11215396807988908812, 0.88164923480869672456, 1.66773656436981165818, 1.37337679382844646270, 0.63028492641384614714, 1.01125376110821929565, 1.33531081175024102237, 1.85257457150729587525, 1.55821480096593045772, 0.89728643877772529258, 1.19609176824570306863, 1.40371261292143079125, 1.52078129870091416365, 1.45483893280140907933, 0.74574760658048411255, 0.89043431182443744110, 1.29575514686297488254, 1.70561930583839815867, 1.63967693993889285231, 0.80556118775908303054, 1.07527231896192132510, 1.19152877946297497402, 1.97761610902409024071, 1.68325633848272460114, 0.66747047180841412306, 1.16623875415628752705, 1.64519035640451916080, 2.16245411616157445778, 1.86809434562020859616, 0.93447198417229304646, 1.35107676129377130003]
-    expected_allMs_shapes_4 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_4 = [np.array([0.00001323000000633893, 0.00734670000000988387, 0.00000000763456877713, 0.00000000456000378356, 0.00000000234534606673]), np.array([0.00001323000000645904, 0.00734670000001055781, 0.00000000763456945101, 0.00000000456000378356, 0.00000000234534666574]), np.array([0.00001323000000633893, 0.00734670000001106001, 0.00000000763456936560, 0.00000000456000456820, 0.00000000234534606673]), np.array([0.00001323000000645904, 0.00734670000001173395, 0.00000000763457003949, 0.00000000456000456820, 0.00000000234534666574]), np.array([0.00001323000000726110, 0.00734670000001027938, 0.00000000763456917234, 0.00000000456000386261, 0.00000000234534659205]), np.array([0.00001323000000738122, 0.00734670000001095332, 0.00000000763456984623, 0.00000000456000386261, 0.00000000234534719106]), np.array([0.00001323000000726110, 0.00734670000001145553, 0.00000000763456976082, 0.00000000456000464724, 0.00000000234534659205]), np.array([0.00001323000000738122, 0.00734670000001212947, 0.00000000763457043471, 0.00000000456000464724, 0.00000000234534719106]), np.array([0.00001323000000699098, 0.00734670000000988387, 0.00000000763456877713, 0.00000000456000410959, 0.00000000234534639275]), np.array([0.00001323000000711110, 0.00734670000001055781, 0.00000000763456945101, 0.00000000456000410959, 0.00000000234534699176]), np.array([0.00001323000000699098, 0.00734670000001106001, 0.00000000763456936560, 0.00000000456000489422, 0.00000000234534639275]), np.array([0.00001323000000711110, 0.00734670000001173395, 0.00000000763457003949, 0.00000000456000489422, 0.00000000234534699176]), np.array([0.00001323000000791315, 0.00734670000001027938, 0.00000000763456917234, 0.00000000456000418863, 0.00000000234534691807]), np.array([0.00001323000000803327, 0.00734670000001095332, 0.00000000763456984623, 0.00000000456000418863, 0.00000000234534751708]), np.array([0.00001323000000791315, 0.00734670000001145553, 0.00000000763456976082, 0.00000000456000497327, 0.00000000234534691807]), np.array([0.00001323000000803327, 0.00734670000001212947, 0.00000000763457043471, 0.00000000456000497327, 0.00000000234534751708])]
-    expected_allIntervals_4 = [np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03945555,  0.03946075], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03945555,  0.03946075], [ 0.41480622,  0.41481102]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03946075,  0.03946555], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03946075,  0.03946555], [ 0.41480622,  0.41481102]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03945555,  0.03946075], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03945555,  0.03946075], [ 0.41480622,  0.41481102]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03946075,  0.03946555], [ 0.41480102,  0.41480622]]),  np.array([[ 0.03945555,  0.73764311], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03946075,  0.03946555], [ 0.41480622,  0.41481102]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03945555,  0.03946075], [ 0.41480102,  0.41480622]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03945555,  0.03946075], [ 0.41480622,  0.41481102]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03946075,  0.03946555], [ 0.41480102,  0.41480622]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.,          1.27667396], [ 0.03946075,  0.03946555], [ 0.41480622,  0.41481102]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03945555,  0.03946075], [ 0.41480102,  0.41480622]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03945555,  0.03946075], [ 0.41480622,  0.41481102]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03946075,  0.03946555], [ 0.41480102,  0.41480622]]),  np.array([[ 0.73764311,  1.3828272 ], [-1.,         -0.99999   ], [ 1.27667396,  1.532344  ], [ 0.03946075,  0.03946555], [ 0.41480622,  0.41481102]])]
-    expected_transforms_4 = [[np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])]]
-    allMs_4, allErrors_4, allIntervals_4 = CS.getSubdivisionIntervals(Ms_4,errors_4,trackedInterval_4,False,level_4)
-    allMs_first_4 = [M[0,0,0,0,0] for Ms in allMs_4 for M in Ms]
-    allMs_shapes_4 = [M.shape for Ms in allMs_4 for M in Ms]
-    allTransforms_4 = [interval.transforms for interval in allIntervals_4]
-    assert np.allclose(expected_allMs_first_4,allMs_first_4)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_4,expected_allMs_shapes_4)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_4,expected_allErrors_4)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_4,expected_allIntervals_4))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_4,expected_transforms_4))
-
-    trackedInterval_5 = CS.TrackedInterval(np.array([[-1.,         -0.48027223]]))
-    Ms_5 = [np.array([ 6.69963798e-01,  2.18239042e-01, -1.22498853e+00, -3.73956662e-02,
-                        9.35209315e-02,  1.63685598e-03, -2.66713529e-03, -3.29041801e-05,
-                        3.98745618e-05,  3.80448135e-07, -3.67357962e-07, -2.85931225e-09,
-                        2.29566748e-09,  1.50932004e-11, -1.03726751e-11, -5.90607284e-14,
-                        3.54786084e-14,  1.76079511e-16, -9.46841952e-17])]
-    errors_5 = np.array([.00000001232332])
-    level_5 = 3
-    expected_allMs_first_5 = [0.90621664678249913560, 1.08962810694742184836]
-    expected_allMs_shapes_5 = [(19,), (19,)]
-    expected_allErrors_5 = [np.array([0.00000001232332948602]),np.array([0.00000001232332948602])]
-    expected_allIntervals_5 = [np.array([[-1.,         -0.72988304]]), np.array([[-0.72988304 -0.48027223]])]
-    expected_transforms_5 = [[np.array([[ 0.51972777], [-0.48027223]])], [np.array([[0.48027223], [0.51972777]])]]
-    allMs_5, allErrors_5, allIntervals_5 = CS.getSubdivisionIntervals(Ms_5,errors_5,trackedInterval_5,False,level_5)
-    allMs_first_5 = [M[0] for Ms in allMs_5 for M in Ms]
-    allMs_shapes_5 = [M.shape for Ms in allMs_5 for M in Ms]
-    allTransforms_5 = [interval.transforms for interval in allIntervals_5]
-    assert np.allclose(expected_allMs_first_5,allMs_first_5)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_5,expected_allMs_shapes_5)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_5,expected_allErrors_5)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_5,expected_allIntervals_5))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_5,expected_transforms_5))
-
-    trackedInterval_6 = CS.TrackedInterval(np.array([[-1.,         -0.48027223]]))
-    Ms_6 = [np.array([ 6.69963798e-01,  2.18239042e-01, -1.22498853e+00, -3.73956662e-02,
-                        9.35209315e-02,  1.63685598e-03, -2.66713529e-03, -3.29041801e-05,
-                        3.98745618e-05,  3.80448135e-07, -3.67357962e-07, -2.85931225e-09,
-                        2.29566748e-09,  1.50932004e-11, -1.03726751e-11, -5.90607284e-14,
-                        3.54786084e-14,  1.76079511e-16, -9.46841952e-17])]
-    errors_6 = np.array([.000000412034002340])
-    level_6 = 7
-    expected_allMs_first_6 = [0.90621664678249913560, 1.08962810694742184836]
-    expected_allMs_shapes_6 = [(19,), (19,)]
-    expected_allErrors_6 = [np.array([0.00000041203401182602]),np.array([0.00000041203401182602])]
-    expected_allIntervals_6 = [np.array([[-1.,         -0.72988304]]), np.array([[-0.72988304, -0.48027223]])]
-    expected_transforms_6 = [[np.array([[ 0.51972777], [-0.48027223]])], [np.array([[0.48027223], [0.51972777]])]]
-    allMs_6, allErrors_6, allIntervals_6 = CS.getSubdivisionIntervals(Ms_6,errors_6,trackedInterval_6,False,level_6)
-    allMs_first_6 = [M[0] for Ms in allMs_6 for M in Ms]
-    allMs_shapes_6 = [M.shape for Ms in allMs_6 for M in Ms]
-    allTransforms_6 = [interval.transforms for interval in allIntervals_6]
-    assert np.allclose(expected_allMs_first_6,allMs_first_6)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_6,expected_allMs_shapes_6)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_6,expected_allErrors_6)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_6,expected_allIntervals_6))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_6,expected_transforms_6))
-
-    trackedInterval_7 = CS.TrackedInterval(np.array([[-1.,         1.       ],
-                                                    [-1.,         1.       ],
-                                                    [-1.,         1.       ],
-                                                    [-1.,         1.       ],
-                                                    [-0.2184735,  1.       ]]))
-    Ms_7 = [np.array([[[[[ 4.08360446e+00, -9.44316964e-01],
-                            [-1.33295089e+00,  2.83295089e+00]],
-
-                            [[-2.40405000e+00, -1.35277734e-16],
-                            [-6.76388669e-17,  6.76388669e-17]],
-
-                            [[ 1.45351250e+00, -6.76388669e-17],
-                            [ 1.20586305e-16,  1.01458300e-16]]],
-
-
-                            [[[ 1.50000000e+00,  0.00000000e+00],
-                            [-3.38194334e-17,  3.38194334e-17]],
-
-                            [[ 1.69097167e-17, -1.69097167e-17],
-                            [ 1.69097167e-17, -1.69097167e-17]],
-
-                            [[ 1.11022302e-16,  0.00000000e+00],
-                            [ 3.38194334e-17, -3.38194334e-17]]],
-
-
-                            [[[ 3.75000000e-01,  0.00000000e+00],
-                            [ 2.77555756e-17,  0.00000000e+00]],
-
-                            [[-2.39100054e-18, -2.53645751e-17],
-                            [ 3.62104340e-17, -8.45485836e-18]],
-
-                            [[-2.53645751e-17,  2.53645751e-17],
-                            [-2.53645751e-17,  2.53645751e-17]]]],
-
-
-
-                        [[[[-1.05000000e-01,  8.45485836e-18],
-                            [-2.39100054e-18, -2.53645751e-17]],
-
-                            [[ 1.53296594e-16, -4.22742918e-17],
-                            [-2.05825740e-17,  7.60937252e-17]],
-
-                            [[ 6.15750090e-17, -3.38194334e-17],
-                            [-4.46652923e-17,  1.69097167e-17]]],
-
-
-                            [[[ 0.00000000e+00,  0.00000000e+00],
-                            [ 0.00000000e+00,  0.00000000e+00]],
-
-                            [[ 9.41125857e-17,  1.69097167e-17],
-                            [ 1.69097167e-17, -1.69097167e-17]],
-
-                            [[-9.41125857e-17, -1.69097167e-17],
-                            [-1.69097167e-17,  1.69097167e-17]]],
-
-
-                            [[[ 1.93007173e-17,  8.45485836e-18],
-                            [ 1.93007173e-17,  8.45485836e-18]],
-
-                            [[ 8.45485836e-18, -8.45485836e-18],
-                            [-3.01465762e-17, -2.53645751e-17]],
-
-                            [[-2.77555756e-17,  0.00000000e+00],
-                            [ 1.08458589e-17,  1.69097167e-17]]]],
-
-
-
-                        [[[[ 5.51250000e-01, -6.76388669e-17],
-                            [-3.38194334e-17,  3.38194334e-17]],
-
-                            [[ 1.31432164e-16,  1.18368017e-16],
-                            [ 1.12304159e-16, -8.45485836e-17]],
-
-                            [[-1.10914381e-18, -1.09913159e-16],
-                            [-1.26822875e-16,  1.26822875e-16]]],
-
-
-                            [[[-1.50732881e-17, -1.26822875e-17],
-                            [-1.50732881e-17, -1.26822875e-17]],
-
-                            [[ 8.32667268e-17,  0.00000000e+00],
-                            [ 4.46652923e-17, -1.69097167e-17]],
-
-                            [[ 8.14302982e-17,  2.95920043e-17],
-                            [ 4.28288637e-17,  1.26822875e-17]]],
-
-
-                            [[[-4.22311136e-17, -6.34114377e-18],
-                            [-3.25807549e-17, -2.11371459e-18]],
-
-                            [[ 2.89510759e-17,  1.26822875e-17],
-                            [ 9.65035863e-18,  4.22742918e-18]],
-
-                            [[ 7.11821894e-17,  1.90234313e-17],
-                            [ 4.22311136e-17,  6.34114377e-18]]]]]), 
-        np.array([[[[[ 8.21242682e+00, -3.35660838e+00,  1.78346906e+00],
-                                [-3.00000000e+00,  0.00000000e+00,  0.00000000e+00],
-                                [ 2.25000000e+00,  0.00000000e+00,  0.00000000e+00]],
-
-                                [[-2.40405000e+00, -1.77327602e-16,  2.47248501e-16],
-                                [ 5.36682823e-16,  1.50055034e-16, -2.06040418e-17],
-                                [-3.74802563e-16,  9.09458333e-17,  6.18121253e-17]],
-
-                                [[ 1.45351250e+00, -2.95546003e-17,  4.12080835e-17],
-                                [ 2.33069664e-17,  5.91092006e-17, -8.24161670e-17],
-                                [-1.16534832e-17, -2.95546003e-17,  4.12080835e-17]]],
-
-
-                                [[[ 5.00000000e-01,  0.00000000e+00,  0.00000000e+00],
-                                [-3.67328042e-17,  2.64307834e-17,  1.03020209e-17],
-                                [-1.83664021e-17,  1.32153917e-17,  5.15101044e-18]],
-
-                                [[-3.67328042e-17,  2.64307834e-17,  1.03020209e-17],
-                                [-7.34656085e-17,  5.28615667e-17,  2.06040418e-17],
-                                [-3.67328042e-17,  2.64307834e-17,  1.03020209e-17]],
-
-                                [[-1.83664021e-17,  1.32153917e-17,  5.15101044e-18],
-                                [-3.67328042e-17,  2.64307834e-17,  1.03020209e-17],
-                                [-1.83664021e-17,  1.32153917e-17,  5.15101044e-18]]],
-
-
-                                [[[ 1.25000000e-01,  0.00000000e+00,  0.00000000e+00],
-                                [ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00],
-                                [ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00]],
-
-                                [[ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00],
-                                [ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00],
-                                [ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00]],
-
-                                [[ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00],
-                                [ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00],
-                                [ 0.00000000e+00,  0.00000000e+00,  0.00000000e+00]]]],
-
-
-
-                            [[[[-1.05000000e-01, -2.74595877e-17,  2.06040418e-17],
-                                [ 1.73167091e-16, -5.35677130e-17,  7.46896513e-17],
-                                [-6.19473131e-17, -4.27881084e-19, -3.47693205e-17]],
-
-                                [[ 1.11354966e-16,  2.88484540e-17,  5.40856096e-17],
-                                [ 3.77152572e-17,  7.44570088e-17, -5.66611148e-17],
-                                [-7.97835766e-17, -1.23596295e-17,  6.43876305e-17]],
-
-                                [[-1.47066405e-16,  2.28790854e-17,  2.70428048e-17],
-                                [-9.49817690e-17,  1.88274939e-16, -1.21048745e-16],
-                                [-3.80712959e-17, -4.20568568e-17,  9.40059405e-17]]],
-
-
-                                [[[ 7.20414539e-18,  7.67390413e-18,  1.28775261e-17],
-                                [ 1.06710302e-17, -7.13331919e-17,  5.15101044e-18],
-                                [ 4.39369496e-17, -1.87568792e-17,  2.57550522e-18]],
-
-                                [[-6.27945782e-17, -1.84716252e-17,  2.57550522e-17],
-                                [ 2.88165816e-17,  3.06956165e-17,  5.15101044e-17],
-                                [ 1.06710302e-17, -7.13331919e-17,  5.15101044e-18]],
-
-                                [[-2.95286588e-17,  3.41046875e-17,  2.31795470e-17],
-                                [-6.27945782e-17, -1.84716252e-17,  2.57550522e-17],
-                                [ 7.20414539e-18,  7.67390413e-18,  1.28775261e-17]]],
-
-
-                                [[[-2.48818456e-17,  1.98978955e-18,  9.01426827e-18],
-                                [-1.11622567e-17,  2.08892958e-17,  1.80285365e-17],
-                                [-6.51544350e-18, -1.12256021e-17,  3.86325783e-18]],
-
-                                [[-1.11622567e-17,  2.08892958e-17,  1.80285365e-17],
-                                [-2.60617740e-17, -4.49024085e-17,  1.54530313e-17],
-                                [ 2.55705475e-17, -5.54148755e-18,  7.72651566e-18]],
-
-                                [[-6.51544350e-18, -1.12256021e-17,  3.86325783e-18],
-                                [ 2.55705475e-17, -5.54148755e-18,  7.72651566e-18],
-                                [ 1.18509586e-17, -2.44409938e-17, -1.28775261e-18]]]],
-
-
-
-                            [[[[ 5.51250000e-01,  1.16235600e-16, -2.06040418e-17],
-                                [-2.46649694e-16, -4.12080835e-17,  1.03020209e-17],
-                                [ 1.13935673e-16,  7.38865007e-18, -1.03020209e-17]],
-
-                                [[ 3.66141733e-16, -1.02314062e-16, -6.95386409e-17],
-                                [ 8.38535004e-17,  1.27889084e-16,  1.03020209e-17],
-                                [ 1.15742016e-17,  1.87568792e-17, -2.57550522e-18]],
-
-                                [[ 6.59450653e-17, -2.27364583e-17, -1.54530313e-17],
-                                [ 1.68622935e-16, -7.19037000e-17, -4.12080835e-17],
-                                [-4.42533475e-17,  5.65558918e-17,  1.54530313e-17]]],
-
-
-                                [[[-5.55111512e-17,  0.00000000e+00,  0.00000000e+00],
-                                [-7.20414539e-18, -7.67390413e-18, -1.28775261e-17],
-                                [ 4.83070058e-17, -7.67390413e-18, -1.28775261e-17]],
-
-                                [[ 3.32659194e-17,  5.25763127e-17, -2.57550522e-18],
-                                [ 5.82674161e-18,  1.47773001e-17, -2.06040418e-17],
-                                [-2.74391778e-17, -3.77990125e-17, -1.80285365e-17]],
-
-                                [[ 4.83070058e-17, -7.67390413e-18, -1.28775261e-17],
-                                [ 1.30308870e-17,  2.24512043e-17, -7.72651566e-18],
-                                [-3.52761188e-17,  3.01251084e-17,  5.15101044e-18]]],
-
-
-                                [[[ 8.97722862e-18, -2.64307834e-17, -1.03020209e-17],
-                                [-3.60207270e-18, -3.83695206e-18, -6.43876305e-18],
-                                [-1.25793013e-17,  2.25938313e-17,  3.86325783e-18]],
-
-                                [[ 1.66329597e-17,  2.62881563e-17, -1.28775261e-18],
-                                [ 2.91337080e-18,  7.38865007e-18, -1.03020209e-17],
-                                [-1.37195889e-17, -1.88995063e-17, -9.01426827e-18]],
-
-                                [[-1.25793013e-17,  2.25938313e-17,  3.86325783e-18],
-                                [ 6.51544350e-18,  1.12256021e-17, -3.86325783e-18],
-                                [ 1.90947448e-17, -1.13682292e-17, -7.72651566e-18]]]]]), 
-        np.array([[[[[ 6.83742682e+00, -3.35660838e+00,  1.78346906e+00],
-          [-1.50000000e+00, -1.18218401e-16,  1.64832334e-16],
-          [ 1.12500000e+00,  1.18218401e-16, -1.64832334e-16]],
-
-         [[-2.40405000e+00, -4.20498680e-17,  2.47248501e-16],
-          [ 5.87724868e-16, -4.22892534e-16, -1.64832334e-16],
-          [-2.93862434e-16,  2.11446267e-16,  8.24161670e-17]],
-
-         [[ 1.45351250e+00,  5.91092006e-17, -8.24161670e-17],
-          [ 4.66139329e-17,  1.18218401e-16, -1.64832334e-16],
-          [-2.33069664e-17, -5.91092006e-17,  8.24161670e-17]]],
-
-
-        [[[ 5.00000000e-01,  1.35277734e-16,  0.00000000e+00],
-          [ 8.68914463e-17, -1.38401551e-16,  5.15101044e-17],
-          [-4.34457231e-17,  6.92007753e-17, -2.57550522e-17]],
-
-         [[-2.57129630e-16,  1.85015484e-16,  7.21141461e-17],
-          [ 2.20396825e-16, -1.58584700e-16, -6.18121253e-17],
-          [-1.10198413e-16,  7.92923501e-17,  3.09060626e-17]],
-
-         [[ 6.71291890e-18, -4.27699920e-17,  3.60570731e-17],
-          [-1.60357055e-16,  1.91263117e-16, -3.09060626e-17],
-          [ 8.01785274e-17, -9.56315587e-17,  1.54530313e-17]]],
-
-
-        [[[ 1.25000000e-01,  8.24161670e-17, -2.06040418e-17],
-          [ 1.23624251e-16, -1.64832334e-16,  4.12080835e-17],
-          [-6.18121253e-17,  8.24161670e-17, -2.06040418e-17]],
-
-         [[-1.46931217e-16,  1.05723133e-16,  4.12080835e-17],
-          [ 2.93862434e-16, -2.11446267e-16, -8.24161670e-17],
-          [-1.46931217e-16,  1.05723133e-16,  4.12080835e-17]],
-
-         [[ 6.18121253e-17, -8.24161670e-17,  2.06040418e-17],
-          [-1.23624251e-16,  1.64832334e-16, -4.12080835e-17],
-          [ 6.18121253e-17, -8.24161670e-17,  2.06040418e-17]]]],
-
-
-
-       [[[[-1.05000000e-01,  2.70292469e-18,  6.69631357e-17],
-          [ 1.52687903e-16,  5.30650240e-17, -1.91875139e-16],
-          [-8.55271524e-17, -1.99248162e-17,  9.85130746e-17]],
-
-         [[ 1.76896583e-16,  4.02845022e-17, -9.01426827e-18],
-          [ 6.52167226e-17, -2.45836208e-17, -1.28775261e-17],
-          [-5.09747634e-17,  2.55072021e-17,  1.15897735e-17]],
-
-         [[-1.02690251e-16, -2.10658324e-17,  5.79488674e-18],
-          [-4.08572472e-17,  4.05697563e-17,  1.41652787e-17],
-          [ 1.12454225e-17, -1.36771823e-17, -4.50713413e-18]]],
-
-
-        [[[-1.42419592e-17, -9.23581259e-19,  1.28775261e-18],
-          [-8.24888592e-18,  2.82779459e-17,  7.72651566e-18],
-          [-1.42419592e-17, -9.23581259e-19,  1.28775261e-18]],
-
-         [[-2.84839183e-17, -1.84716252e-18,  2.57550522e-18],
-          [-1.64977718e-17,  5.65558918e-17,  1.54530313e-17],
-          [-2.84839183e-17, -1.84716252e-18,  2.57550522e-18]],
-
-         [[-1.42419592e-17, -9.23581259e-19,  1.28775261e-18],
-          [-8.24888592e-18,  2.82779459e-17,  7.72651566e-18],
-          [-1.42419592e-17, -9.23581259e-19,  1.28775261e-18]]],
-
-
-        [[[-7.12097958e-18, -4.61790629e-19,  6.43876305e-19],
-          [-4.12444296e-18,  1.41389729e-17,  3.86325783e-18],
-          [-7.12097958e-18, -4.61790629e-19,  6.43876305e-19]],
-
-         [[-1.42419592e-17, -9.23581259e-19,  1.28775261e-18],
-          [-8.24888592e-18,  2.82779459e-17,  7.72651566e-18],
-          [-1.42419592e-17, -9.23581259e-19,  1.28775261e-18]],
-
-         [[-7.12097958e-18, -4.61790629e-19,  6.43876305e-19],
-          [-4.12444296e-18,  1.41389729e-17,  3.86325783e-18],
-          [-7.12097958e-18, -4.61790629e-19,  6.43876305e-19]]]],
-
-
-
-       [[[[ 5.51250000e-01,  1.94386934e-16, -8.24161670e-17],
-          [ 2.91337080e-18,  7.38865007e-18, -1.03020209e-17],
-          [ 1.87783470e-17,  2.64307834e-17,  1.03020209e-17]],
-
-         [[ 2.14080244e-16, -2.20911423e-17,  1.13322230e-16],
-          [ 4.82900118e-17, -1.55311267e-16,  5.15101044e-17],
-          [-3.90997348e-18,  1.07780742e-16, -2.06040418e-17]],
-
-         [[ 1.53659813e-17, -2.42305477e-17, -6.05243726e-17],
-          [ 2.86424816e-17,  2.34426046e-17, -7.98406618e-17],
-          [-4.20372458e-18,  3.34125189e-18,  4.24958361e-17]]],
-
-
-        [[[ 1.60789184e-16, -1.08422564e-16, -3.15499389e-17],
-          [-9.10640271e-17,  7.33229814e-17,  3.86325783e-18],
-          [ 3.03557392e-17, -5.92553220e-17, -5.79488674e-18]],
-
-         [[ 7.02754807e-17, -1.70523437e-17, -1.15897735e-17],
-          [ 4.68503204e-17, -1.13682292e-17, -7.72651566e-18],
-          [-2.34251602e-17,  5.68411458e-18,  3.86325783e-18]],
-
-         [[-1.70480952e-17,  3.85086532e-17, -6.43876305e-19],
-          [-9.01686943e-18,  2.10319228e-17,  2.96183100e-17],
-          [ 1.96847090e-17,  1.20778699e-17, -1.09458972e-17]]],
-
-
-        [[[ 7.78652131e-17, -5.79769204e-17, -1.64188458e-17],
-          [-4.55320135e-17,  3.66614907e-17,  1.93162891e-18],
-          [ 1.77072487e-17, -2.58620225e-17, -2.25356707e-18]],
-
-         [[ 3.51377403e-17, -8.52617187e-18, -5.79488674e-18],
-          [ 2.34251602e-17, -5.68411458e-18, -3.86325783e-18],
-          [-1.17125801e-17,  2.84205729e-18,  1.93162891e-18]],
-
-         [[-5.99466856e-18,  2.30199652e-17,  3.21938152e-19],
-          [-4.50843471e-18,  1.05159614e-17,  1.48091550e-17],
-          [ 7.31297546e-18,  2.27329639e-18, -6.11682489e-18]]]]]), 
-          
-          np.array([[[[[-4.90571520e-02,  2.59687165e+00],
-          [ 1.33295089e+00, -2.83295089e+00],
-          [-9.99713170e-01,  2.12471317e+00]],
-
-         [[-2.88486000e-01,  6.76388669e-17],
-          [ 2.42554313e-17, -1.35277734e-16],
-          [-1.21277156e-17,  6.76388669e-17]],
-
-         [[ 1.74421500e-01,  6.76388669e-17],
-          [ 1.44841736e-16, -3.38194334e-17],
-          [-1.49069165e-16,  3.80468626e-17]]],
-
-
-        [[[ 5.00000000e-01,  0.00000000e+00],
-          [-1.93007173e-17, -8.45485836e-18],
-          [ 1.93007173e-17,  8.45485836e-18]],
-
-         [[-9.65035863e-18, -4.22742918e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 9.65035863e-18,  4.22742918e-18]],
-
-         [[ 2.89510759e-17,  1.26822875e-17],
-          [ 1.93007173e-17,  8.45485836e-18],
-          [-9.65035863e-18, -4.22742918e-18]]],
-
-
-        [[[ 1.25000000e-01, -1.26822875e-17],
-          [-1.93007173e-17, -8.45485836e-18],
-          [ 9.65035863e-18,  4.22742918e-18]],
-
-         [[-2.11371459e-18,  2.11371459e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 2.11371459e-18, -2.11371459e-18]],
-
-         [[ 2.89510759e-17,  1.26822875e-17],
-          [ 1.93007173e-17,  8.45485836e-18],
-          [-9.65035863e-18, -4.22742918e-18]]]],
-
-
-
-       [[[[-1.05000000e-01, -4.22742918e-17],
-          [-6.76388669e-17,  6.76388669e-17],
-          [ 3.67285727e-18, -5.91840085e-17]],
-
-         [[-4.34266138e-17, -1.90234313e-17],
-          [ 7.72028690e-17,  3.38194334e-17],
-          [-3.37762552e-17, -1.47960021e-17]],
-
-         [[ 2.04314502e-17,  9.40602992e-17],
-          [ 5.80748647e-17, -1.69097167e-16],
-          [-1.30344585e-18,  1.08856301e-16]]],
-
-
-        [[[-4.82517931e-18, -2.11371459e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 4.82517931e-18,  2.11371459e-18]],
-
-         [[ 9.65035863e-18,  4.22742918e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [-9.65035863e-18, -4.22742918e-18]],
-
-         [[-4.82517931e-18, -2.11371459e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 4.82517931e-18,  2.11371459e-18]]],
-
-
-        [[[ 2.41258966e-18,  1.05685729e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [-2.41258966e-18, -1.05685729e-18]],
-
-         [[-4.82517931e-18, -2.11371459e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 4.82517931e-18,  2.11371459e-18]],
-
-         [[ 2.41258966e-18,  1.05685729e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [-2.41258966e-18, -1.05685729e-18]]]],
-
-
-
-       [[[[ 5.51250000e-01,  0.00000000e+00],
-          [-4.78200108e-18, -5.07291501e-17],
-          [ 1.69097167e-17, -1.69097167e-17]],
-
-         [[ 7.60937252e-17, -7.60937252e-17],
-          [-1.18368017e-16,  1.18368017e-16],
-          [ 7.60937252e-17, -7.60937252e-17]],
-
-         [[ 4.88927215e-17, -2.11371459e-17],
-          [-3.74922907e-17,  9.30034419e-17],
-          [ 4.88927215e-17, -2.11371459e-17]]],
-
-
-        [[[-4.22742918e-18,  4.22742918e-18],
-          [ 2.04962175e-17,  2.11371459e-17],
-          [ 2.47236467e-17,  1.69097167e-17]],
-
-         [[-9.65035863e-18, -4.22742918e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 9.65035863e-18,  4.22742918e-18]],
-
-         [[ 4.22742918e-18, -4.22742918e-18],
-          [-1.19550027e-18, -1.26822875e-17],
-          [-5.42292945e-18, -8.45485836e-18]]],
-
-
-        [[[ 6.93889390e-18,  0.00000000e+00],
-          [ 2.35281464e-17,  4.22742918e-18],
-          [ 1.65892525e-17,  4.22742918e-18]],
-
-         [[-9.65035863e-18, -4.22742918e-18],
-          [ 0.00000000e+00,  0.00000000e+00],
-          [ 9.65035863e-18,  4.22742918e-18]],
-
-         [[-6.93889390e-18,  0.00000000e+00],
-          [-4.22742918e-18,  4.22742918e-18],
-          [ 2.71146472e-18,  4.22742918e-18]]]]]), 
-          
-          np.array([[[[[ 2.53662057e+00, -3.35660838e+00,  1.78346906e+00],
-          [ 2.92580625e+00, -2.70555467e-16,  0.00000000e+00]],
-
-         [[ 1.20202500e+00, -3.67748935e-16,  4.12080835e-17],
-          [-3.60607500e+00, -5.91092006e-17,  8.24161670e-17]],
-
-         [[-7.26756250e-01,  3.00110068e-16, -4.12080835e-17],
-          [ 2.18026875e+00, -2.87614800e-16, -1.64832334e-16]]],
-
-
-        [[[ 5.00000000e-01,  3.90756669e-17, -3.09060626e-17],
-          [ 2.81512551e-17, -2.70012915e-17, -5.66611148e-17]],
-
-         [[ 7.34656085e-17, -5.28615667e-17, -2.06040418e-17],
-          [-7.34656085e-17,  5.28615667e-17,  2.06040418e-17]],
-
-         [[-2.81512551e-17,  2.70012915e-17,  5.66611148e-17],
-          [ 4.53143534e-17, -2.58602752e-17,  3.60570731e-17]]],
-
-
-        [[[ 1.25000000e-01, -3.69432504e-18,  5.15101044e-18],
-          [ 1.64977718e-17, -5.65558918e-17, -1.54530313e-17]],
-
-         [[ 7.22009189e-17, -5.47443860e-17, -2.09259799e-17],
-          [-7.22009189e-17,  5.47443860e-17,  2.09259799e-17]],
-
-         [[-1.64977718e-17,  5.65558918e-17,  1.54530313e-17],
-          [ 5.69678366e-17,  3.69432504e-18, -5.15101044e-18]]]],
-
-
-
-       [[[[-1.05000000e-01, -1.35006457e-17, -2.83305574e-17],
-          [-3.13972891e-17, -9.23581259e-18,  1.28775261e-17]],
-
-         [[ 1.53197857e-16, -1.52764947e-17, -1.99601654e-17],
-          [ 5.44384576e-17, -7.13135139e-20, -5.79488674e-18]],
-
-         [[-5.85092832e-17,  3.76207287e-17,  3.54131968e-18],
-          [-9.12958369e-18,  3.00181381e-17, -3.54131968e-18]]],
-
-
-        [[[ 5.87292322e-17,  2.13940542e-19,  1.73846602e-17],
-          [-7.03827154e-17, -2.97685408e-17,  2.38234233e-17]],
-
-         [[-8.86418828e-17,  3.02677354e-17,  1.67407839e-17],
-          [ 1.29111948e-16,  2.99824814e-17, -6.43876305e-18]],
-
-         [[ 4.00301668e-17, -1.54191218e-17, -3.15499389e-17],
-          [-6.88467484e-17, -1.52764947e-17, -1.99601654e-17]]],
-
-
-        [[[ 3.94821323e-17,  1.51695245e-17,  1.12678353e-17],
-          [-2.50738415e-17,  1.78283785e-19,  1.44872169e-17]],
-
-         [[-6.45559738e-17, -1.49912407e-17,  3.21938152e-18],
-          [ 4.43209414e-17, -1.51338677e-17, -8.37039196e-18]],
-
-         [[ 3.01325996e-17,  7.35299331e-18, -1.31994642e-17],
-          [-2.43058580e-17,  7.42430683e-18, -7.40457750e-18]]]],
-
-
-
-       [[[[ 5.51250000e-01,  0.00000000e+00,  0.00000000e+00],
-          [-2.29960828e-16,  5.71263999e-17,  6.18121253e-17]],
-
-         [[ 9.62020668e-17, -5.49939833e-17, -4.12080835e-17],
-          [ 2.78600497e-16, -3.59518500e-17, -2.06040418e-17]],
-
-         [[ 4.32878885e-17, -9.80632070e-18, -3.34815678e-17],
-          [ 8.06534884e-17, -1.11835129e-16, -7.98406618e-17]]],
-
-
-        [[[ 3.15898405e-18,  3.26105981e-17, -2.18917944e-17],
-          [ 1.86120154e-17,  1.20065564e-17, -1.67407839e-17]],
-
-         [[ 6.23033517e-17, -3.19722709e-17, -2.57550522e-18],
-          [-4.20683194e-17,  6.20973793e-17,  7.72651566e-18]],
-
-         [[-8.49449917e-18,  3.05599783e-18,  1.93162891e-17],
-          [ 6.95853215e-18, -1.75480439e-17,  2.44672996e-17]]],
-
-
-        [[[ 1.57949203e-18,  1.63052991e-17, -1.09458972e-17],
-          [ 9.30600768e-18,  6.00327818e-18, -8.37039196e-18]],
-
-         [[ 3.11516759e-17, -1.59861355e-17, -1.28775261e-18],
-          [-2.10341597e-17,  3.10486897e-17,  3.86325783e-18]],
-
-         [[-4.24724958e-18,  1.52799892e-18,  9.65814457e-18],
-          [ 3.47926607e-18, -8.77402196e-18,  1.22336498e-17]]]]])]
-    level_7 = 1
-    errors_7 = np.array([.000000203420340234,.000745704567,.00000000456456757,.045674567,.00000008474646])
-    expected_allMs_first_7 = [5.67627456479412817458, 10.57502137862851121497, 8.78174328527667391597, -2.32697349062049596569, 0.96054435083336220380, 3.37136997013587169647, 7.28878074701508804623, 5.49550265366325252359, 1.05987235659383105535, -2.32569628078006074290, 2.98273604413587190010, 7.66379636072424830218, 7.32613077632454245958, 0.28685875190987780714, 5.03271481677655252440, 3.51078233947761564337, 4.37755572911082335708, 4.03989014471112017901, 0.92458543073545929047, 1.74647418516312935566, 3.32957369642231881812, 8.22832051025670097033, 6.43504241690486544769, -2.60857759482511308846, 3.82447766044020731613, 1.02466910176406234001, 4.94207987864327868976, 3.14880178529144272304, 0.77826825238921393257, 0.53823702882678348125, 0.63603517576406254364, 5.31709549235243805754, 4.97942990795273310312, 0.00525464770526046232, 4.37659682382568160364, 1.16408147110580650896, 2.03085486073901311244, 1.69318927633931015642, 0.64298132653084194565, 1.09035619221225754671, 7.19107039514341739306, 11.07995332207827487991, 9.28667522872643758092, -1.82204154717073252279, 1.46547629428312542466, 4.88616580048516091495, 7.79371269046485082299, 6.00043459711301530035, 1.56480430004359427620, -1.82076433733029752204, 4.49753187448516111857, 8.16872830417401019076, 7.83106271977430523634, 0.79179069535964075044, 5.53764676022631530117, 5.02557816982690486185, 4.88248767256058613384, 4.54482208816088295578, 1.42951737418522228928, 2.25140612861289213242, 4.84436952677160803660, 8.73325245370646463527, 6.93997436035462733628, -2.10364565137534986761, 4.32940960388997009289, 2.53946493211335200257, 5.44701182209304146653, 3.65373372874120549980, 1.28320019583897693138, 1.04316897227654648006, 2.15083100611335176211, 5.82202743580220083430, 5.48436185140249587988, 0.51018659115502340562, 4.88152876727544438040, 2.67887730145509594948, 2.53578680418877677738, 2.19812121978907315523, 1.14791326998060494446, 1.59528813566202054552, 5.59302443540758353890, 10.49177124924196569111, 8.69849315589012839212, -2.41022362000704060137, 0.87729422144681712403, 3.28811984074932706079, 7.20553061762854341055, 5.41225252427670788791, 0.97662222720728641967, -2.40894641016660537858, 2.89948591474932726442, 7.58054623133770366650, 7.24288064693799782390, 0.20360862252333294942, 4.94946468739000700054, 3.42753221009107100770, 4.29430559972427872140, 3.95664001532457554333, 0.84133530134891443275, 1.66322405577658427589, 3.24632356703577418244, 8.14507038087015544647, 6.35179228751832081201, -2.69182772421165772414, 3.74122753105366268045, 0.94141897237751781535, 4.85882974925673405409, 3.06555165590489764327, 0.69501812300266929689, 0.45498689944023879006, 0.55278504637751790796, 5.23384536296589342186, 4.89617977856618846744, -0.07799548168128433989, 4.29334669443913696796, 1.08083134171926187328, 1.94760473135246847676, 1.60993914695276574278, 0.55973119714429719895, 1.00710606282571291104, 7.10782026575687275738, 10.99670319269172935606, 9.20342509933989205706, -1.90529167655727738051, 1.38222616489658034489, 4.80291567109861627927, 7.71046256107830618731, 5.91718446772647066467, 1.48155417065704941848, -1.90401446671684237977, 4.41428174509861648289, 8.08547817478746466691, 7.74781259038776060066, 0.70854056597309589272, 5.45439663083977066549, 4.94232804044036022617, 4.79923754317404149816, 4.46157195877433832010, 1.34626724479867743156, 2.16815599922634749674, 4.76111939738506251274, 8.65000232431991911142, 6.85672423096808270060, -2.18689578076189450329, 4.24615947450342545721, 2.45621480272680692281, 5.36376169270649683085, 3.57048359935466086412, 1.19995006645243229570, 0.95991884289000206643, 2.06758087672680668234, 5.73877730641565619862, 5.40111172201595124420, 0.42693646176847860341, 4.79827863788889974472, 2.59562717206855086971, 2.45253667480223169761, 2.11487109040252896364, 1.06466314059406008674, 1.51203800627547613189]
-    expected_allMs_shapes_7 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_7 = [np.array([0.00000020342038181112, 0.00074570456707858843, 0.00000000456463283559, 0.04567456700003586639, 0.00000008474651875765]), np.array([0.00000020342038019132, 0.00074570456706070083, 0.00000000456461494799, 0.04567456700003586639, 0.00000008474650305208]), np.array([0.00000020342038181112, 0.00074570456706659803, 0.00000000456462684039, 0.04567456700001940734, 0.00000008474651875765]), np.array([0.00000020342038019132, 0.00074570456704871042, 0.00000000456460895279, 0.04567456700001940734, 0.00000008474650305208]), np.array([0.00000020342037113500, 0.00074570456707218275, 0.00000000456462642992, 0.04567456700003583864, 0.00000008474650584443]), np.array([0.00000020342036951519, 0.00074570456705429515, 0.00000000456460854231, 0.04567456700003583864, 0.00000008474648944746]), np.array([0.00000020342037113500, 0.00074570456706019234, 0.00000000456462043471, 0.04567456700001838732, 0.00000008474650584443]), np.array([0.00000020342036951519, 0.00074570456704230474, 0.00000000456460254711, 0.04567456700001838732, 0.00000008474648944746]), np.array([0.00000020342038521452, 0.00074570456707907470, 0.00000000456463332179, 0.04567456700003555414, 0.00000008474651956799]), np.array([0.00000020342038359472, 0.00074570456706118710, 0.00000000456461543419, 0.04567456700003555414, 0.00000008474650317102]), np.array([0.00000020342038521452, 0.00074570456706708429, 0.00000000456462732659, 0.04567456700002021919, 0.00000008474651956799]), np.array([0.00000020342038359472, 0.00074570456704919669, 0.00000000456460943899, 0.04567456700002021919, 0.00000008474650317102]), np.array([0.00000020342037453840, 0.00074570456707266902, 0.00000000456462691612, 0.04567456700003552639, 0.00000008474650665476]), np.array([0.00000020342037291859, 0.00074570456705478141, 0.00000000456460902851, 0.04567456700003552639, 0.00000008474649025779]), np.array([0.00000020342037453840, 0.00074570456706067861, 0.00000000456462092091, 0.04567456700001919917, 0.00000008474650665476]), np.array([0.00000020342037291859, 0.00074570456704279101, 0.00000000456460303331, 0.04567456700001919917, 0.00000008474649025779]), np.array([0.00000020342038162461, 0.00074570456707858843, 0.00000000456463283559, 0.04567456700003584558, 0.00000008474651866440]), np.array([0.00000020342038000480, 0.00074570456706070083, 0.00000000456461494799, 0.04567456700003584558, 0.00000008474650303276]), np.array([0.00000020342038162461, 0.00074570456706659803, 0.00000000456462684039, 0.04567456700001931713, 0.00000008474651866440]), np.array([0.00000020342038000480, 0.00074570456704871042, 0.00000000456460895279, 0.04567456700001931713, 0.00000008474650303276]), np.array([0.00000020342037094848, 0.00074570456707218275, 0.00000000456462642992, 0.04567456700003581782, 0.00000008474650575117]), np.array([0.00000020342036932867, 0.00074570456705429515, 0.00000000456460854231, 0.04567456700003581782, 0.00000008474648935420]), np.array([0.00000020342037094848, 0.00074570456706019234, 0.00000000456462043471, 0.04567456700001829711, 0.00000008474650575117]), np.array([0.00000020342036932867, 0.00074570456704230474, 0.00000000456460254711, 0.04567456700001829711, 0.00000008474648935420]), np.array([0.00000020342038502801, 0.00074570456707907470, 0.00000000456463332179, 0.04567456700003553333, 0.00000008474651947473]), np.array([0.00000020342038340820, 0.00074570456706118710, 0.00000000456461543419, 0.04567456700003553333, 0.00000008474650307776]), np.array([0.00000020342038502801, 0.00074570456706708429, 0.00000000456462732659, 0.04567456700002012898, 0.00000008474651947473]), np.array([0.00000020342038340820, 0.00074570456704919669, 0.00000000456460943899, 0.04567456700002012898, 0.00000008474650307776]), np.array([0.00000020342037435188, 0.00074570456707266902, 0.00000000456462691612, 0.04567456700003550557, 0.00000008474650656150]), np.array([0.00000020342037273207, 0.00074570456705478141, 0.00000000456460902851, 0.04567456700003550557, 0.00000008474649016454]), np.array([0.00000020342037435188, 0.00074570456706067861, 0.00000000456462092091, 0.04567456700001910896, 0.00000008474650656150]), np.array([0.00000020342037273207, 0.00074570456704279101, 0.00000000456460303331, 0.04567456700001910896, 0.00000008474649016454])]
-    expected_allIntervals_7 = [np.array([[-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.03945555,  1.        ], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [-0.2184735,   0.41480102]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-1.,          0.03945555], [ 0.41480102,  1.        ]]),  np.array([[ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [ 0.03945555,  1.        ], [-0.2184735,   0.41480102]]),  np.array([[0.03945555, 1.        ], [0.03945555, 1.        ], [0.03945555, 1.        ], [0.03945555, 1.        ], [0.41480102, 1.        ]])]
-    expected_transforms_7 = [[np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])]]
-    allMs_7, allErrors_7, allIntervals_7 = CS.getSubdivisionIntervals(Ms_7,errors_7,trackedInterval_7,False,level_7)
-    allMs_first_7 = [M[0,0,0,0,0] for Ms in allMs_7 for M in Ms]
-    allMs_shapes_7 = [M.shape for Ms in allMs_7 for M in Ms]
-    allTransforms_7 = [interval.transforms for interval in allIntervals_7]
-    assert np.allclose(expected_allMs_first_7,allMs_first_7)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_7,expected_allMs_shapes_7)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_7,expected_allErrors_7)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_7,expected_allIntervals_7))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_7,expected_transforms_7))
-
-    trackedInterval_8 = CS.TrackedInterval(np.array([[ 0.08600977,  0.1526489 ],
-                                                    [-1.,         -1+1e-5],
-                                                    [-0.95417107, -0.80126711]]))
-    Ms_8 = [np.array([[[ 7.35038766e-03,  4.44916068e-04,  7.95637666e-03],
-                        [-5.50186312e-04,  1.19708171e-18, -4.83542451e-20]],
-
-                    [[ 2.87993052e-02,  5.76240555e-18, -4.13686254e-19],
-                        [ 2.73339051e-02,  3.42010626e-20, -4.16235200e-21]],
-
-                    [[-5.88967146e-05, -7.67655018e-20, -4.24202506e-22],
-                        [-5.80164717e-20,  2.02102879e-21, -3.20168454e-23]],
-
-                    [[ 4.87676378e-04, -1.49632791e-20,  3.79126356e-22],
-                        [ 4.50516959e-21,  1.62221708e-22, -4.93937912e-24]]]), 
-            np.array([[[-1.68459073e-06,  1.26145762e-01],
-                        [ 2.70612965e-05, -3.85321532e-16],
-                        [ 1.54635980e-05, -6.47641707e-17],
-                        [ 5.79884923e-06,  1.73832165e-18],
-                        [ 1.28863316e-06, -1.53108143e-20],
-                        [ 1.28863316e-07,  4.35035965e-23]],
-
-                    [[ 1.24830555e-01, -9.30887954e-16],
-                        [-1.95839123e-15,  1.69259248e-17],
-                        [ 2.01239919e-17,  2.41329377e-18],
-                        [ 7.76587403e-19, -6.52985838e-20],
-                        [-1.35646803e-20,  5.82334349e-22],
-                        [ 5.41754243e-23, -1.69761760e-24]],
-
-                    [[ 9.77319850e-04, -1.31472753e-17],
-                        [ 3.39298863e-17, -2.81038274e-18],
-                        [-2.65085503e-18,  3.73793918e-19],
-                        [ 6.03680985e-20, -9.76938402e-21],
-                        [-5.32276430e-22,  9.00334315e-23],
-                        [ 1.58077877e-24, -2.70045420e-25]],
-
-                    [[-3.92482564e-05,  3.71489858e-19],
-                        [ 1.11831865e-18, -8.85658375e-21],
-                        [-2.69184037e-20, -1.45083768e-21],
-                        [ 2.77359921e-23,  4.35197323e-23],
-                        [ 2.86353295e-24, -4.11904944e-25],
-                        [-1.46906132e-26,  1.24451738e-27]],
-
-                    [[ 2.43737461e-04,  3.20039943e-21],
-                        [ 1.99298700e-21,  5.49466463e-22],
-                        [ 5.43601526e-22, -1.04848894e-22],
-                        [-1.93723652e-23,  2.89574122e-24],
-                        [ 2.02924150e-25, -2.73040469e-26],
-                        [-6.56165771e-28,  8.29187586e-29]]]), 
-            np.array([[[ 7.82899240e-03, -4.44916068e-04, -7.95637666e-03],
-                        [ 1.06348094e-02,  6.06694601e-18, -1.39879477e-19],
-                        [ 2.65870235e-03, -3.97829518e-20,  9.10079548e-22]],
-
-                    [[-6.28491250e-04,  1.27987287e-18, -1.15758419e-20],
-                        [ 2.92792557e-18, -1.97762413e-20, -1.73813116e-21],
-                        [-1.41696049e-20,  2.03066147e-22,  7.80132009e-24]],
-
-                    [[ 7.80604670e-03,  5.94908870e-20, -1.29545887e-21],
-                        [ 3.82176670e-20, -6.63258736e-21,  1.44429585e-22],
-                        [-2.52242272e-22,  4.37760605e-23, -9.53256688e-25]]])]
-    level_8 = 6
-    errors_8 = np.array([.00000420340234,.00045704567,.000056456757])
-    expected_allMs_first_8 = [-0.00870879253848562407, -0.12075753163810644397, 0.00838489276915926375, -0.00794995327246854865, 0.00538823036189399474, 0.00762605350314218312, 0.01984548944854796354, 0.00411672876558271533, 0.00806439336628413973, 0.02060432871456504764, 0.13026249076558221729, 0.00730555410026706344]
-    expected_allMs_shapes_8 = [(4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3)]
-    expected_allErrors_8 = [np.array([0.00000420340234010364, 0.00045704567000039154, 0.00005645675700005311]), np.array([0.00000420340234010364, 0.00045704567000039154, 0.00005645675700005269]),np.array([0.00000420340234011262, 0.00045704567000039208, 0.00005645675700005311]),np.array([0.00000420340234011262, 0.00045704567000039208, 0.00005645675700005269])]
-    expected_allIntervals_8 = [np.array([[ 0.08600977,  0.12064398], [-1.,         -0.99999   ], [-0.95417107, -0.87470264]]), np.array([[ 0.08600977,  0.12064398], [-1.,         -0.99999   ], [-0.87470264, -0.80126711]]), np.array([[ 0.12064398,  0.1526489 ], [-1.,         -0.99999   ], [-0.95417107, -0.87470264]]), np.array([[ 0.12064398,  0.1526489 ], [-1.,         -0.99999   ], [-0.87470264, -0.80126711]])]
-    expected_transforms_8 = [[np.array([[ 0.51972777,  1.        ,  1.        ], [-0.48027223,  0.        ,  0.        ]]), np.array([[ 1.        ,  1.        ,  0.51972777], [ 0.        ,  0.        , -0.48027223]])], [np.array([[ 0.51972777,  1.        ,  1.        ], [-0.48027223,  0.        ,  0.        ]]), np.array([[1.        , 1.        , 0.48027223], [0.        , 0.        , 0.51972777]])], [np.array([[0.48027223, 1.        , 1.        ], [0.51972777, 0.        , 0.        ]]), np.array([[ 1.        ,  1.        ,  0.51972777], [ 0.        ,  0.        , -0.48027223]])], [np.array([[0.48027223, 1.        , 1.        ], [0.51972777, 0.        , 0.        ]]), np.array([[1.        , 1.        , 0.48027223], [0.        , 0.        , 0.51972777]])]]
-    allMs_8, allErrors_8, allIntervals_8 = CS.getSubdivisionIntervals(Ms_8,errors_8,trackedInterval_8,False,level_8)
-    allMs_first_8 = [M[0,0,0] for Ms in allMs_8 for M in Ms]
-    allMs_shapes_8 = [M.shape for Ms in allMs_8 for M in Ms]
-    allTransforms_8 = [interval.transforms for interval in allIntervals_8]
-    assert np.allclose(expected_allMs_first_8,allMs_first_8)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_8,expected_allMs_shapes_8)])
-    assert np.all([np.allclose(error,exp) for error,exp in zip(allErrors_8,expected_allErrors_8)])
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_8,expected_allIntervals_8))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_8,expected_transforms_8))
-
-    trackedInterval_9 = CS.TrackedInterval(np.array([[ 0.08600977,  0.1526489 ],
-                                                    [-1.,         -1+1e-5],
-                                                    [-0.95417107, -0.80126711]]))
-    Ms_9 = [np.array([[[ 7.35038766e-03,  4.44916068e-04,  7.95637666e-03],
-                        [-5.50186312e-04,  1.19708171e-18, -4.83542451e-20]],
-
-                    [[ 2.87993052e-02,  5.76240555e-18, -4.13686254e-19],
-                        [ 2.73339051e-02,  3.42010626e-20, -4.16235200e-21]],
-
-                    [[-5.88967146e-05, -7.67655018e-20, -4.24202506e-22],
-                        [-5.80164717e-20,  2.02102879e-21, -3.20168454e-23]],
-
-                    [[ 4.87676378e-04, -1.49632791e-20,  3.79126356e-22],
-                        [ 4.50516959e-21,  1.62221708e-22, -4.93937912e-24]]]), 
-            np.array([[[-1.68459073e-06,  1.26145762e-01],
-                        [ 2.70612965e-05, -3.85321532e-16],
-                        [ 1.54635980e-05, -6.47641707e-17],
-                        [ 5.79884923e-06,  1.73832165e-18],
-                        [ 1.28863316e-06, -1.53108143e-20],
-                        [ 1.28863316e-07,  4.35035965e-23]],
-
-                    [[ 1.24830555e-01, -9.30887954e-16],
-                        [-1.95839123e-15,  1.69259248e-17],
-                        [ 2.01239919e-17,  2.41329377e-18],
-                        [ 7.76587403e-19, -6.52985838e-20],
-                        [-1.35646803e-20,  5.82334349e-22],
-                        [ 5.41754243e-23, -1.69761760e-24]],
-
-                    [[ 9.77319850e-04, -1.31472753e-17],
-                        [ 3.39298863e-17, -2.81038274e-18],
-                        [-2.65085503e-18,  3.73793918e-19],
-                        [ 6.03680985e-20, -9.76938402e-21],
-                        [-5.32276430e-22,  9.00334315e-23],
-                        [ 1.58077877e-24, -2.70045420e-25]],
-
-                    [[-3.92482564e-05,  3.71489858e-19],
-                        [ 1.11831865e-18, -8.85658375e-21],
-                        [-2.69184037e-20, -1.45083768e-21],
-                        [ 2.77359921e-23,  4.35197323e-23],
-                        [ 2.86353295e-24, -4.11904944e-25],
-                        [-1.46906132e-26,  1.24451738e-27]],
-
-                    [[ 2.43737461e-04,  3.20039943e-21],
-                        [ 1.99298700e-21,  5.49466463e-22],
-                        [ 5.43601526e-22, -1.04848894e-22],
-                        [-1.93723652e-23,  2.89574122e-24],
-                        [ 2.02924150e-25, -2.73040469e-26],
-                        [-6.56165771e-28,  8.29187586e-29]]]), 
-            np.array([[[ 7.82899240e-03, -4.44916068e-04, -7.95637666e-03],
-                        [ 1.06348094e-02,  6.06694601e-18, -1.39879477e-19],
-                        [ 2.65870235e-03, -3.97829518e-20,  9.10079548e-22]],
-
-                    [[-6.28491250e-04,  1.27987287e-18, -1.15758419e-20],
-                        [ 2.92792557e-18, -1.97762413e-20, -1.73813116e-21],
-                        [-1.41696049e-20,  2.03066147e-22,  7.80132009e-24]],
-
-                    [[ 7.80604670e-03,  5.94908870e-20, -1.29545887e-21],
-                        [ 3.82176670e-20, -6.63258736e-21,  1.44429585e-22],
-                        [-2.52242272e-22,  4.37760605e-23, -9.53256688e-25]]])]
-    level_9 = 4
-    errors_9 = np.array([.00000420340234,.00045704567,.000056456757])
-    expected_allMs_first_9 = [-0.00870879253848562407, -0.12075753163810644397, 0.00838489276915926375, -0.00794995327246854865, 0.00538823036189399474, 0.00762605350314218312, 0.01984548944854796354, 0.00411672876558271533, 0.00806439336628413973, 0.02060432871456504764, 0.13026249076558221729, 0.00730555410026706344]
-    expected_allMs_shapes_9 = [(4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3)]
-    expected_allErrors_flattened_9 = [0.00000420340234010364, 0.00045704567000039154, 0.00005645675700005311, 0.00000420340234010364, 0.00045704567000039154, 0.00005645675700005269, 0.00000420340234011262, 0.00045704567000039208, 0.00005645675700005311, 0.00000420340234011262, 0.00045704567000039208, 0.00005645675700005269]
-    expected_allIntervals_9 = [np.array([[ 0.08600977,  0.12064398], [-1.,         -0.99999   ], [-0.95417107, -0.87470264]]),  np.array([[ 0.08600977,  0.12064398], [-1.,         -0.99999   ], [-0.87470264, -0.80126711]]),  np.array([[ 0.12064398,  0.1526489 ], [-1.,         -0.99999   ], [-0.95417107, -0.87470264]]),  np.array([[ 0.12064398,  0.1526489 ], [-1.,         -0.99999   ], [-0.87470264, -0.80126711]])]
-    expected_transforms_9 = [[np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])]]
-    allMs_9, allErrors_9, allIntervals_9 = CS.getSubdivisionIntervals(Ms_9,errors_9,trackedInterval_9,False,level_9)
-    allMs_first_9 = [M[0,0,0] for Ms in allMs_9 for M in Ms]
-    allMs_shapes_9 = [M.shape for Ms in allMs_9 for M in Ms]
-    allErrors_flattened_9 = [item for error_arr in allErrors_9 for item in error_arr]
-    allTransforms_9 = [interval.transforms for interval in allIntervals_9]
-    assert np.allclose(expected_allMs_first_9,allMs_first_9)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_9,expected_allMs_shapes_9)])
-    assert np.allclose(expected_allErrors_flattened_9,allErrors_flattened_9)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_9,expected_allIntervals_9))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_9,expected_transforms_9))
-    
-    trackedInterval_10 = CS.TrackedInterval(np.array([[ 0.1110769,   0.11932933],
-                                                    [-1.,         -0.99371914],
-                                                    [-0.87726429, -0.85851422]]))
-    Ms_10 = [np.array([[[-4.97165769e-05,  5.56342937e-04,  1.19642019e-04],       
-                        [-4.80456420e-04,  1.74630072e-20, -8.77767068e-23]],       
-
-                    [[ 4.31081351e-04,  8.50677180e-20, -7.65165445e-22],        
-                        [ 4.13282291e-04,  5.67814700e-23, -9.39634614e-25]],       
-
-                    [[-6.46024106e-06, -1.27250950e-22, -1.57047985e-25],        
-                        [-1.14389590e-22,  4.33016340e-25, -7.98122860e-28]],       
-
-                    [[ 9.26169059e-07, -3.47197103e-24,  1.09509057e-26],        
-                        [ 1.05055527e-24,  4.54034949e-27, -1.72223167e-29]]]), 
-                        
-            np.array([[[-2.71749444e-06,  1.54688085e-02],
-                        [ 7.34199307e-10, -2.15348756e-18],
-                        [ 4.19538527e-10, -1.36905583e-19],
-                        [ 1.57326955e-10,  4.16353877e-22],
-                        [ 3.49615456e-11, -4.32036135e-25],
-                        [ 3.49615456e-12,  1.46292238e-28]],
-
-                    [[ 1.54688354e-02, -1.42784526e-17],
-                        [-3.09612230e-17,  1.87639173e-20],
-                        [ 3.27003415e-20,  5.79265649e-22],
-                        [ 1.87780208e-22, -1.77989846e-24],
-                        [-3.77182628e-25,  1.86468052e-27],
-                        [ 1.78843638e-28, -6.45664292e-31]],
-
-                    [[ 2.08863674e-06, -2.02373910e-20],
-                        [ 7.89842350e-20, -9.61118664e-22],
-                        [-6.63831902e-22,  1.20054036e-23],
-                        [ 1.75485288e-24, -3.57815789e-26],
-                        [-1.83063430e-27,  3.87876291e-29],
-                        [ 6.48720224e-31, -1.38403444e-32]],
-
-                    [[-5.33127238e-07,  8.75223696e-23],
-                        [ 2.81139376e-22, -1.25450538e-25],
-                        [-7.88904432e-25, -5.43901299e-27],
-                        [ 1.15596284e-28,  1.83886021e-29],
-                        [ 1.15419495e-30, -2.04443871e-32],
-                        [-7.15741834e-34,  7.34426478e-36]],
-
-                    [[ 5.73236294e-08,  7.66929515e-26],
-                        [ 2.32276984e-27,  3.29616462e-27],
-                        [ 2.21964840e-27, -5.18052559e-29],
-                        [-8.73949675e-30,  1.62174287e-31],
-                        [ 1.07177769e-32, -1.79646466e-34],
-                        [-4.11881198e-36,  6.48798307e-38]]]), 
-            np.array([[[ 1.19168975e-04, -5.56342937e-04, -1.19642019e-04],
-                        [ 1.58531014e-04,  9.19590565e-20, -2.62536469e-22],        
-                        [ 3.96327536e-05, -7.19887572e-23,  2.03988284e-25]],       
-
-                    [[-5.56677977e-04,  1.91367880e-20, -1.73942480e-23],        
-                        [ 4.47221870e-20, -3.35095740e-23, -4.18052662e-25],        
-                        [-2.58982619e-23,  4.20228466e-26,  2.29665581e-28]],       
-
-                    [[ 1.19711768e-04,  1.21496768e-22, -3.28105373e-25],        
-                        [ 7.13200296e-23, -1.54072377e-24,  4.16076705e-27],        
-                        [-5.61702813e-26,  1.21344436e-27, -3.27693997e-30]]])]
-    level_10 = 7
-    errors_10 = np.array([.0000002534523452345,.000006020304,.0000000996789])
-    expected_allMs_first_10 = [-0.00022806388487373632, -0.01486187460336119842, 0.00056692145138122934, 0.00033299959350165142, 0.00060693389663880968, 0.00000585797300584141, -0.00090700831082173418, -0.01486187393167794080, 0.00072701619737733803, -0.00034594483244634649, 0.00060693456832206383, 0.00016595271900195021, 0.00000381364600708721, 0.00060730513985275333, 0.00001496676774160653, 0.00056487712438247500, 0.01607611363985275016, -0.00054609671063378135, -0.00026184848894091067, 0.00060730581153597713, 0.00017506151373771527, 0.00029921498943447718, 0.01607611431153596962, -0.00038600196463767249]
-    expected_allMs_shapes_10 = [(4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3)]
-    expected_allErrors_flattened_10 = [0.00000025345234523853, 0.00000602030400008932, 0.00000009967890000295, 0.00000025345234523839, 0.00000602030400008932, 0.00000009967890000298, 0.00000025345234523853, 0.00000602030400008932, 0.00000009967890000310, 0.00000025345234523839, 0.00000602030400008932, 0.00000009967890000292, 0.00000025345234523778, 0.00000602030400008932, 0.00000009967890000295, 0.00000025345234523801, 0.00000602030400008932, 0.00000009967890000298, 0.00000025345234523778, 0.00000602030400008932, 0.00000009967890000310, 0.00000025345234523801, 0.00000602030400008932, 0.00000009967890000292]
-    expected_allIntervals_10 = [np.array([[ 0.1110769,   0.11536592], [-1.,         -0.99673566], [-0.87726429, -0.86751936]]),  np.array([[ 0.1110769,   0.11536592], [-1.,         -0.99673566], [-0.86751936, -0.85851422]]),  np.array([[ 0.1110769,   0.11536592], [-0.99673566, -0.99371914], [-0.87726429, -0.86751936]]),  np.array([[ 0.1110769,   0.11536592], [-0.99673566, -0.99371914], [-0.86751936, -0.85851422]]),  np.array([[ 0.11536592,  0.11932933], [-1.,         -0.99673566], [-0.87726429, -0.86751936]]),  np.array([[ 0.11536592,  0.11932933], [-1.,         -0.99673566], [-0.86751936, -0.85851422]]),  np.array([[ 0.11536592,  0.11932933], [-0.99673566, -0.99371914], [-0.87726429, -0.86751936]]),  np.array([[ 0.11536592,  0.11932933], [-0.99673566, -0.99371914], [-0.86751936, -0.85851422]])]
-    expected_transforms_10 = [[np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])]]
-    allMs_10, allErrors_10, allIntervals_10 = CS.getSubdivisionIntervals(Ms_10,errors_10,trackedInterval_10,False,level_10)
-    allMs_first_10 = [M[0,0,0] for Ms in allMs_10 for M in Ms]
-    allMs_shapes_10 = [M.shape for Ms in allMs_10 for M in Ms]
-    allErrors_flattened_10 = [item for error_arr in allErrors_10 for item in error_arr]
-    allTransforms_10 = [interval.transforms for interval in allIntervals_10]
-    assert np.allclose(expected_allMs_first_10,allMs_first_10)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_10,expected_allMs_shapes_10)])
-    assert np.allclose(expected_allErrors_flattened_10,allErrors_flattened_10)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_10,expected_allIntervals_10))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_10,expected_transforms_10))
-
-    trackedInterval_11 = CS.TrackedInterval(np.array([[ 0.08600977,  0.1526489 ],
-                                                    [-1.,         -.9],
-                                                    [-0.95417107, -0.80126711]]))
-    Ms_11 = [np.array([[[ 7.35038766e-03,  4.44916068e-04,  7.95637666e-03],
-                        [-5.50186312e-04,  1.19708171e-18, -4.83542451e-20]],
-
-                    [[ 2.87993052e-02,  5.76240555e-18, -4.13686254e-19],
-                        [ 2.73339051e-02,  3.42010626e-20, -4.16235200e-21]],
-
-                    [[-5.88967146e-05, -7.67655018e-20, -4.24202506e-22],
-                        [-5.80164717e-20,  2.02102879e-21, -3.20168454e-23]],
-
-                    [[ 4.87676378e-04, -1.49632791e-20,  3.79126356e-22],
-                        [ 4.50516959e-21,  1.62221708e-22, -4.93937912e-24]]]), 
-            np.array([[[-1.68459073e-06,  1.26145762e-01],
-                        [ 2.70612965e-05, -3.85321532e-16],
-                        [ 1.54635980e-05, -6.47641707e-17],
-                        [ 5.79884923e-06,  1.73832165e-18],
-                        [ 1.28863316e-06, -1.53108143e-20],
-                        [ 1.28863316e-07,  4.35035965e-23]],
-
-                    [[ 1.24830555e-01, -9.30887954e-16],
-                        [-1.95839123e-15,  1.69259248e-17],
-                        [ 2.01239919e-17,  2.41329377e-18],
-                        [ 7.76587403e-19, -6.52985838e-20],
-                        [-1.35646803e-20,  5.82334349e-22],
-                        [ 5.41754243e-23, -1.69761760e-24]],
-
-                    [[ 9.77319850e-04, -1.31472753e-17],
-                        [ 3.39298863e-17, -2.81038274e-18],
-                        [-2.65085503e-18,  3.73793918e-19],
-                        [ 6.03680985e-20, -9.76938402e-21],
-                        [-5.32276430e-22,  9.00334315e-23],
-                        [ 1.58077877e-24, -2.70045420e-25]],
-
-                    [[-3.92482564e-05,  3.71489858e-19],
-                        [ 1.11831865e-18, -8.85658375e-21],
-                        [-2.69184037e-20, -1.45083768e-21],
-                        [ 2.77359921e-23,  4.35197323e-23],
-                        [ 2.86353295e-24, -4.11904944e-25],
-                        [-1.46906132e-26,  1.24451738e-27]],
-
-                    [[ 2.43737461e-04,  3.20039943e-21],
-                        [ 1.99298700e-21,  5.49466463e-22],
-                        [ 5.43601526e-22, -1.04848894e-22],
-                        [-1.93723652e-23,  2.89574122e-24],
-                        [ 2.02924150e-25, -2.73040469e-26],
-                        [-6.56165771e-28,  8.29187586e-29]]]), 
-            np.array([[[ 7.82899240e-03, -4.44916068e-04, -7.95637666e-03],
-                        [ 1.06348094e-02,  6.06694601e-18, -1.39879477e-19],
-                        [ 2.65870235e-03, -3.97829518e-20,  9.10079548e-22]],
-
-                    [[-6.28491250e-04,  1.27987287e-18, -1.15758419e-20],
-                        [ 2.92792557e-18, -1.97762413e-20, -1.73813116e-21],
-                        [-1.41696049e-20,  2.03066147e-22,  7.80132009e-24]],
-
-                    [[ 7.80604670e-03,  5.94908870e-20, -1.29545887e-21],
-                        [ 3.82176670e-20, -6.63258736e-21,  1.44429585e-22],
-                        [-2.52242272e-22,  4.37760605e-23, -9.53256688e-25]]])]
-    level_11 = 4
-    errors_11 = np.array([.0000002534523452345,.000006020304,.0000000996789])
-    expected_allMs_first_11 = [-0.00213967620798279426, -0.12077315270137434222, 0.00256326749903206755, -0.00138083694196571884, 0.00537260929862631159, 0.00180442823301498433, -0.01581757797312523250, -0.12074839565868808466, 0.01330297745615168122, -0.01505873870710815535, 0.00539736634131220139, 0.01254413819013460406, 0.01328689032590835498, 0.00410110770231576077, 0.00224276809615694440, 0.01404572959192543907, 0.13024686970231547090, 0.00148392883013986183, 0.02694289366076592016, 0.00412586474500004075, 0.01298247805327656067, 0.02770173292678300425, 0.13027162674499939699, 0.01222363878725948351]
-    expected_allMs_shapes_11 = [(4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3), (4, 2, 3), (5, 6, 2), (3, 3, 3)]
-    expected_allErrors_flattened_11 = [0.00000025345234536572, 0.00000602030400072763, 0.00000009967890007059, 0.00000025345234536544, 0.00000602030400072763, 0.00000009967890006976, 0.00000025345234536572, 0.00000602030400072767, 0.00000009967890008093, 0.00000025345234536544, 0.00000602030400072767, 0.00000009967890008010, 0.00000025345234537860, 0.00000602030400072815, 0.00000009967890007059, 0.00000025345234537899, 0.00000602030400072815, 0.00000009967890006976, 0.00000025345234537860, 0.00000602030400072821, 0.00000009967890008093, 0.00000025345234537899, 0.00000602030400072821, 0.00000009967890008010]
-    expected_allIntervals_11 = [np.array([[ 0.08600977,  0.12064398], [-1.,         -0.94802722], [-0.95417107, -0.87470264]]), np.array([[ 0.08600977,  0.12064398], [-1.,         -0.94802722], [-0.87470264, -0.80126711]]), np.array([[ 0.08600977,  0.12064398], [-0.94802722, -0.9       ], [-0.95417107, -0.87470264]]), np.array([[ 0.08600977,  0.12064398], [-0.94802722, -0.9       ], [-0.87470264, -0.80126711]]), np.array([[ 0.12064398,  0.1526489 ], [-1.,         -0.94802722], [-0.95417107, -0.87470264]]), np.array([[ 0.12064398,  0.1526489 ], [-1.,         -0.94802722], [-0.87470264, -0.80126711]]), np.array([[ 0.12064398,  0.1526489 ], [-0.94802722, -0.9       ], [-0.95417107, -0.87470264]]), np.array([[ 0.12064398,  0.1526489 ], [-0.94802722, -0.9       ], [-0.87470264 ,-0.80126711]])]
-    expected_transforms_11 = [[np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.], [-0.48027223,  0.,  0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[ 1.,  0.51972777,  1.], [ 0., -0.48027223,  0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  0.51972777], [ 0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1.], [0.51972777, 0., 0.]]), np.array([[1., 0.48027223, 1.], [0., 0.51972777, 0.]]), np.array([[1., 1., 0.48027223], [0., 0., 0.51972777]])]]
-    allMs_11, allErrors_11, allIntervals_11 = CS.getSubdivisionIntervals(Ms_11,errors_11,trackedInterval_11,False,level_11)
-    allMs_first_11 = [M[0,0,0] for Ms in allMs_11 for M in Ms]
-    allMs_shapes_11 = [M.shape for Ms in allMs_11 for M in Ms]
-    allErrors_flattened_11 = [item for error_arr in allErrors_11 for item in error_arr]
-    allTransforms_11 = [interval.transforms for interval in allIntervals_11]
-    assert np.allclose(expected_allMs_first_11,allMs_first_11)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_11,expected_allMs_shapes_11)])
-    assert np.allclose(expected_allErrors_flattened_11,allErrors_flattened_11)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_11,expected_allIntervals_11))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_11,expected_transforms_11))
-
-    trackedInterval_12 = CS.TrackedInterval(np.array([[1.,  1. + 1e-5]]))
-    Ms_12 = [np.array([ 1.98388230e-01, -7.45867440e-02,  4.04311356e-01,  5.88271449e-01,
-                        2.26026393e-01,  1.32940050e+00, -4.45466673e-01, -8.61593362e-01,
-                        1.63812136e-01,  2.12669974e-01, -2.95468574e-02, -2.95366346e-02,
-                        3.27547093e-03,  2.68367335e-03, -2.48957499e-04, -1.73443093e-04,
-                        1.38658587e-05,  8.41868996e-06, -5.92171480e-07, -3.18960845e-07,
-                        2.00473230e-08,  9.70991815e-09, -5.51888146e-10, -2.42949605e-10,
-                        1.26081368e-11,  5.08808273e-12, -2.43249865e-13, -9.04276654e-14,
-                        4.28823643e-15,  1.47104551e-15])                                 ]
-    level_12 = 1
-    errors_12 = np.array([.00000034523452345])
-    expected_allMs_first_12 = [0.11067191064975603898, 0.50951097496170472123]
-    expected_allMs_shapes_12 = [(30,), (30,)]
-    expected_allErrors_flattened_12 = [0.00000034523455389242, 0.00000034523455389242]
-    expected_allIntervals_12 = [np.array([[1.,        1.0000052]]), np.array([[1.0000052, 1.00001  ]])]
-    expected_transforms_12 = [[np.array([[ 0.51972777], [-0.48027223]])], [np.array([[0.48027223], [0.51972777]])]]
-    allMs_12, allErrors_12, allIntervals_12 = CS.getSubdivisionIntervals(Ms_12,errors_12,trackedInterval_12,False,level_12)
-    allMs_first_12 = [M[0] for Ms in allMs_12 for M in Ms]
-    allMs_shapes_12 = [M.shape for Ms in allMs_12 for M in Ms]
-    allErrors_flattened_12 = [item for error_arr in allErrors_12 for item in error_arr]
-    allTransforms_12 = [interval.transforms for interval in allIntervals_12]
-    assert np.allclose(expected_allMs_first_12,allMs_first_12)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_12,expected_allMs_shapes_12)])
-    assert np.allclose(expected_allErrors_flattened_12,allErrors_flattened_12)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_12,expected_allIntervals_12))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_12,expected_transforms_12))
-
-    trackedInterval_13 = CS.TrackedInterval(np.array([[ 0.03945555,  1.        ],
-                                                        [-1.,          0.03945555],
-                                                        [ 0.03945555,  1.        ],
-                                                        [ 0.03945555,  0.8650092 ],
-                                                        [ 0.41480102,  1.        ]]))
-    Ms_13 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-    level_13 = 2
-    errors_13 = np.array([.00000034523452345,.0000000052342354,.00000013346534,.086745466453,.0000034000003456])
-    expected_allMs_first_13 = [0.36705493861818166978, 0.63198896511584889968, 0.56604659921634359332, 0.12964927225512884457, 0.34840317201161813410, 0.25909747255972598312, 0.81682697225333278368, 0.75088460635382747732, 0.18946285343372770704, 0.53324117914910196259, 0.15487110515972590807, 1.08882377543902464367, 0.79446400489765922615, 0.05137213748305902161, 0.43234097217743211461, 0.60853268210127020588, 1.27366178257650863870, 0.97930201203514322117, 0.31837364984693805603, 0.61717897931491594310, 0.67693448327245975271, 0.94186850977012681607, 0.87592614387062184278, 0.16683481764969676497, 0.31152152289365009352, 0.56897701721400406605, 1.12670651690761092212, 1.06076415100810561576, 0.22664839882829562745, 0.49635953003113386650, 0.46475064981400404651, 1.39870332009330256007, 1.10434354955193736458, 0.08855768287762671997, 0.58732596522550017948, 0.91841222675554834431, 1.58354132723078677714, 1.28918155668942135961, 0.35555919524150575439, 0.77216397236298395246, 0.77622698133280754185, 0.76837964568344552241, 0.70243727978393999400, 0.26603995282272518974, 0.48479385257921447927, 0.66826951527435174416, 0.95321765282092929539, 0.88727528692142398903, 0.32585353400132405222, 0.66963185971669836327, 0.56404314787435172462, 1.22521445600662115538, 0.93085468546525573785, 0.18776281805065536679, 0.56873165274502845978, 1.01770472481589591141, 1.41005246314410515041, 1.11569269260273973288, 0.45476433041453440120, 0.75356965988251234378, 1.08610652598708545824, 1.07825919033772321676, 1.01231682443821835449, 0.30322549821729305464, 0.44791220346124649421, 0.97814905992862977158, 1.26309719747520743383, 1.19715483157570212747, 0.36303907939589191711, 0.63275021059873037821, 0.87392269252862975204, 1.53509400066089907178, 1.24073423011953387629, 0.22494836344522306515, 0.72371664579309658016, 1.32758426947017404984, 1.71993200779838328884, 1.42557223725701787131, 0.49194987580910209957, 0.90855465293058046417, 0.87203505747029730344, 1.13696908396796425578, 1.07102671806845894942, 0.63462939110724447822, 0.85338329086373376775, 0.76407759141184150575, 1.32180709110544825080, 1.25586472520594294444, 0.69444297228584339621, 1.03822129800121731869, 0.65985122401184148622, 1.59380389429113988875, 1.29944412374977469327, 0.55635225633517459976, 0.93732109102954763724, 1.11351280095338567300, 1.77864190142862410582, 1.48428213088725868829, 0.82335376869905352315, 1.12215909816703129920, 1.18191460212457544188, 1.44684862862224239421, 1.38090626272273730990, 0.67181493650181245414, 0.81650164174576578269, 1.07395713606611953317, 1.63168663575972638924, 1.56574426986022108288, 0.73162851768041137213, 1.00133964888324955567, 0.96973076866611962465, 1.90368343894541847128, 1.60932366840405283170, 0.59353780172974235363, 1.09230608407761597967, 1.42339234560766381144, 2.08852144608290268835, 1.79416167554153682673, 0.86053931409362149907, 1.27714409121509953060, 1.28120710018492323101, 1.27335976453556098953, 1.20741739863605546113, 0.77102007167484087891, 0.98977397143132994639, 1.17324963412646732230, 1.45819777167304476251, 1.39225540577353945615, 0.83083365285343979689, 1.17461197856881383039, 1.06902326672646741379, 1.73019457485873640046, 1.43583480431737120497, 0.69274293690277088942, 1.07371177159714381588, 1.52268484366801160057, 1.91503258199622061753, 1.62067281145485520000, 0.95974444926665003486, 1.25854977873462781091, 1.59108664483920114741, 1.58323930918983890592, 1.51729694329033382161, 0.80820561706940874380, 0.95289232231336218337, 1.48312917878074523870, 1.76807731632732290095, 1.70213495042781759459, 0.86801919824800766179, 1.13773032945084606737, 1.37890281138074533018, 2.04007411951301520503, 1.74571434897164934341, 0.72992848229733864329, 1.22869676464521226933, 1.83256438832228951696, 2.22491212665049920005, 1.93055235610913333844, 0.99692999466121778873, 1.41353477178269604231]
-    expected_allMs_shapes_13 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_flattened_13 = [0.00000034523453088437, 0.00000000523424697607, 0.00000013346535034039, 0.08674546645300459935, 0.00000340000035295813, 0.00000034523453100449, 0.00000000523424787459, 0.00000013346535123890, 0.08674546645300459935, 0.00000340000035378177, 0.00000034523453088437, 0.00000000523424874150, 0.00000013346535122310, 0.08674546645300566794, 0.00000340000035295813, 0.00000034523453100449, 0.00000000523424964001, 0.00000013346535212162, 0.08674546645300566794, 0.00000340000035378177, 0.00000034523453220177, 0.00000000523424776651, 0.00000013346535113082, 0.08674546645300472425, 0.00000340000035379865, 0.00000034523453232189, 0.00000000523424866503, 0.00000013346535202934, 0.08674546645300472425, 0.00000340000035462229, 0.00000034523453220177, 0.00000000523424953193, 0.00000013346535201353, 0.08674546645300579284, 0.00000340000035379865, 0.00000034523453232189, 0.00000000523425043045, 0.00000013346535291205, 0.08674546645300579284, 0.00000340000035462229, 0.00000034523453180369, 0.00000000523424710740, 0.00000013346535047172, 0.08674546645300480752, 0.00000340000035317702, 0.00000034523453192381, 0.00000000523424800592, 0.00000013346535137023, 0.08674546645300480752, 0.00000340000035400066, 0.00000034523453180369, 0.00000000523424887283, 0.00000013346535135443, 0.08674546645300587611, 0.00000340000035317702, 0.00000034523453192381, 0.00000000523424977135, 0.00000013346535225295, 0.08674546645300587611, 0.00000340000035400066, 0.00000034523453312109, 0.00000000523424789784, 0.00000013346535126215, 0.08674546645300496017, 0.00000340000035401753, 0.00000034523453324120, 0.00000000523424879636, 0.00000013346535216067, 0.08674546645300496017, 0.00000340000035484117, 0.00000034523453312109, 0.00000000523424966327, 0.00000013346535214487, 0.08674546645300602876, 0.00000340000035401753, 0.00000034523453324120, 0.00000000523425056178, 0.00000013346535304338, 0.08674546645300602876, 0.00000340000035484117, 0.00000034523453153643, 0.00000000523424697607, 0.00000013346535034039, 0.08674546645300491854, 0.00000340000035328416, 0.00000034523453165654, 0.00000000523424787459, 0.00000013346535123890, 0.08674546645300491854, 0.00000340000035410780, 0.00000034523453153643, 0.00000000523424874150, 0.00000013346535122310, 0.08674546645300600101, 0.00000340000035328416, 0.00000034523453165654, 0.00000000523424964001, 0.00000013346535212162, 0.08674546645300600101, 0.00000340000035410780, 0.00000034523453285382, 0.00000000523424776651, 0.00000013346535113082, 0.08674546645300505732, 0.00000340000035412467, 0.00000034523453297394, 0.00000000523424866503, 0.00000013346535202934, 0.08674546645300505732, 0.00000340000035494831, 0.00000034523453285382, 0.00000000523424953193, 0.00000013346535201353, 0.08674546645300612591, 0.00000340000035412467, 0.00000034523453297394, 0.00000000523425043045, 0.00000013346535291205, 0.08674546645300612591, 0.00000340000035494831, 0.00000034523453245574, 0.00000000523424710740, 0.00000013346535047172, 0.08674546645300514058, 0.00000340000035350304, 0.00000034523453257586, 0.00000000523424800592, 0.00000013346535137023, 0.08674546645300514058, 0.00000340000035432668, 0.00000034523453245574, 0.00000000523424887283, 0.00000013346535135443, 0.08674546645300620917, 0.00000340000035350304, 0.00000034523453257586, 0.00000000523424977135, 0.00000013346535225295, 0.08674546645300620917, 0.00000340000035432668, 0.00000034523453377314, 0.00000000523424789784, 0.00000013346535126215, 0.08674546645300527936, 0.00000340000035434356, 0.00000034523453389326, 0.00000000523424879636, 0.00000013346535216067, 0.08674546645300527936, 0.00000340000035516720, 0.00000034523453377314, 0.00000000523424966327, 0.00000013346535214487, 0.08674546645300634795, 0.00000340000035434356, 0.00000034523453389326, 0.00000000523425056178, 0.00000013346535304338, 0.08674546645300634795, 0.00000340000035516720]
-    expected_allIntervals_13 = [[[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]], [[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]]
-    expected_transforms_13 = [[np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])]]
-    allMs_13, allErrors_13, allIntervals_13 = CS.getSubdivisionIntervals(Ms_13,errors_13,trackedInterval_13,False,level_13)
-    allMs_first_13 = [M[0,0,0,0,0] for Ms in allMs_13 for M in Ms]
-    allMs_shapes_13 = [M.shape for Ms in allMs_13 for M in Ms]
-    allErrors_flattened_13 = [item for error_arr in allErrors_13 for item in error_arr]
-    allTransforms_13 = [interval.transforms for interval in allIntervals_13]
-    assert np.allclose(expected_allMs_first_13,allMs_first_13)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_13,expected_allMs_shapes_13)])
-    assert np.allclose(expected_allErrors_flattened_13,allErrors_flattened_13)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_13,expected_allIntervals_13))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_13,expected_transforms_13))
-
-    trackedInterval_14 = CS.TrackedInterval(np.array([[ 0.03945555,  1.        ],
-                            [-1.,          0.03945555],
-                            [ 0.03945555,  1.        ],
-                            [ 0.03945555,  0.8650092 ],
-                            [ 0.41480102,  1.        ]]))
-    Ms_14 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-    level_14 = 6
-    errors_14 = np.array([.00000034523452345,.000000000000035634,.0000000034563456,.00000002112341243,.00000024])
-    expected_allMs_first_14 = [0.36705493861818166978, 0.63198896511584889968, 0.56604659921634359332, 0.12964927225512884457, 0.34840317201161813410, 0.25909747255972598312, 0.81682697225333278368, 0.75088460635382747732, 0.18946285343372770704, 0.53324117914910196259, 0.15487110515972590807, 1.08882377543902464367, 0.79446400489765922615, 0.05137213748305902161, 0.43234097217743211461, 0.60853268210127020588, 1.27366178257650863870, 0.97930201203514322117, 0.31837364984693805603, 0.61717897931491594310, 0.67693448327245975271, 0.94186850977012681607, 0.87592614387062184278, 0.16683481764969676497, 0.31152152289365009352, 0.56897701721400406605, 1.12670651690761092212, 1.06076415100810561576, 0.22664839882829562745, 0.49635953003113386650, 0.46475064981400404651, 1.39870332009330256007, 1.10434354955193736458, 0.08855768287762671997, 0.58732596522550017948, 0.91841222675554834431, 1.58354132723078677714, 1.28918155668942135961, 0.35555919524150575439, 0.77216397236298395246, 0.77622698133280754185, 0.76837964568344552241, 0.70243727978393999400, 0.26603995282272518974, 0.48479385257921447927, 0.66826951527435174416, 0.95321765282092929539, 0.88727528692142398903, 0.32585353400132405222, 0.66963185971669836327, 0.56404314787435172462, 1.22521445600662115538, 0.93085468546525573785, 0.18776281805065536679, 0.56873165274502845978, 1.01770472481589591141, 1.41005246314410515041, 1.11569269260273973288, 0.45476433041453440120, 0.75356965988251234378, 1.08610652598708545824, 1.07825919033772321676, 1.01231682443821835449, 0.30322549821729305464, 0.44791220346124649421, 0.97814905992862977158, 1.26309719747520743383, 1.19715483157570212747, 0.36303907939589191711, 0.63275021059873037821, 0.87392269252862975204, 1.53509400066089907178, 1.24073423011953387629, 0.22494836344522306515, 0.72371664579309658016, 1.32758426947017404984, 1.71993200779838328884, 1.42557223725701787131, 0.49194987580910209957, 0.90855465293058046417, 0.87203505747029730344, 1.13696908396796425578, 1.07102671806845894942, 0.63462939110724447822, 0.85338329086373376775, 0.76407759141184150575, 1.32180709110544825080, 1.25586472520594294444, 0.69444297228584339621, 1.03822129800121731869, 0.65985122401184148622, 1.59380389429113988875, 1.29944412374977469327, 0.55635225633517459976, 0.93732109102954763724, 1.11351280095338567300, 1.77864190142862410582, 1.48428213088725868829, 0.82335376869905352315, 1.12215909816703129920, 1.18191460212457544188, 1.44684862862224239421, 1.38090626272273730990, 0.67181493650181245414, 0.81650164174576578269, 1.07395713606611953317, 1.63168663575972638924, 1.56574426986022108288, 0.73162851768041137213, 1.00133964888324955567, 0.96973076866611962465, 1.90368343894541847128, 1.60932366840405283170, 0.59353780172974235363, 1.09230608407761597967, 1.42339234560766381144, 2.08852144608290268835, 1.79416167554153682673, 0.86053931409362149907, 1.27714409121509953060, 1.28120710018492323101, 1.27335976453556098953, 1.20741739863605546113, 0.77102007167484087891, 0.98977397143132994639, 1.17324963412646732230, 1.45819777167304476251, 1.39225540577353945615, 0.83083365285343979689, 1.17461197856881383039, 1.06902326672646741379, 1.73019457485873640046, 1.43583480431737120497, 0.69274293690277088942, 1.07371177159714381588, 1.52268484366801160057, 1.91503258199622061753, 1.62067281145485520000, 0.95974444926665003486, 1.25854977873462781091, 1.59108664483920114741, 1.58323930918983890592, 1.51729694329033382161, 0.80820561706940874380, 0.95289232231336218337, 1.48312917878074523870, 1.76807731632732290095, 1.70213495042781759459, 0.86801919824800766179, 1.13773032945084606737, 1.37890281138074533018, 2.04007411951301520503, 1.74571434897164934341, 0.72992848229733864329, 1.22869676464521226933, 1.83256438832228951696, 2.22491212665049920005, 1.93055235610913333844, 0.99692999466121778873, 1.41353477178269604231]    # expected_allMs_shapes_14 = [(3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2), (3, 3, 3, 2, 2)]
-    expected_allMs_shapes_14 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_flattened_14 = [0.00000034523453088437, 0.00000000000004721007, 0.00000000345635594039, 0.00000002112341700936, 0.00000024000000735813, 0.00000034523453100449, 0.00000000000004810859, 0.00000000345635683890, 0.00000002112341700936, 0.00000024000000818177, 0.00000034523453088437, 0.00000000000004897550, 0.00000000345635682310, 0.00000002112341808823, 0.00000024000000735813, 0.00000034523453100449, 0.00000000000004987401, 0.00000000345635772162, 0.00000002112341808823, 0.00000024000000818177, 0.00000034523453220177, 0.00000000000004800051, 0.00000000345635673082, 0.00000002112341713583, 0.00000024000000819865, 0.00000034523453232189, 0.00000000000004889903, 0.00000000345635762934, 0.00000002112341713583, 0.00000024000000902229, 0.00000034523453220177, 0.00000000000004976593, 0.00000000345635761353, 0.00000002112341821470, 0.00000024000000819865, 0.00000034523453232189, 0.00000000000005066445, 0.00000000345635851205, 0.00000002112341821470, 0.00000024000000902229, 0.00000034523453180369, 0.00000000000004734140, 0.00000000345635607172, 0.00000002112341722824, 0.00000024000000757702, 0.00000034523453192381, 0.00000000000004823992, 0.00000000345635697023, 0.00000002112341722824, 0.00000024000000840066, 0.00000034523453180369, 0.00000000000004910683, 0.00000000345635695443, 0.00000002112341830711, 0.00000024000000757702, 0.00000034523453192381, 0.00000000000005000535, 0.00000000345635785295, 0.00000002112341830711, 0.00000024000000840066, 0.00000034523453312109, 0.00000000000004813184, 0.00000000345635686215, 0.00000002112341735471, 0.00000024000000841753, 0.00000034523453324120, 0.00000000000004903036, 0.00000000345635776067, 0.00000002112341735471, 0.00000024000000924117, 0.00000034523453312109, 0.00000000000004989727, 0.00000000345635774487, 0.00000002112341843358, 0.00000024000000841753, 0.00000034523453324120, 0.00000000000005079578, 0.00000000345635864338, 0.00000002112341843358, 0.00000024000000924117, 0.00000034523453153643, 0.00000000000004721007, 0.00000000345635594039, 0.00000002112341733539, 0.00000024000000768416, 0.00000034523453165654, 0.00000000000004810859, 0.00000000345635683890, 0.00000002112341733539, 0.00000024000000850780, 0.00000034523453153643, 0.00000000000004897550, 0.00000000345635682310, 0.00000002112341841426, 0.00000024000000768416, 0.00000034523453165654, 0.00000000000004987401, 0.00000000345635772162, 0.00000002112341841426, 0.00000024000000850780, 0.00000034523453285382, 0.00000000000004800051, 0.00000000345635673082, 0.00000002112341746186, 0.00000024000000852467, 0.00000034523453297394, 0.00000000000004889903, 0.00000000345635762934, 0.00000002112341746186, 0.00000024000000934831, 0.00000034523453285382, 0.00000000000004976593, 0.00000000345635761353, 0.00000002112341854073, 0.00000024000000852467, 0.00000034523453297394, 0.00000000000005066445, 0.00000000345635851205, 0.00000002112341854073, 0.00000024000000934831, 0.00000034523453245574, 0.00000000000004734140, 0.00000000345635607172, 0.00000002112341755427, 0.00000024000000790304, 0.00000034523453257586, 0.00000000000004823992, 0.00000000345635697023, 0.00000002112341755427, 0.00000024000000872668, 0.00000034523453245574, 0.00000000000004910683, 0.00000000345635695443, 0.00000002112341863314, 0.00000024000000790304, 0.00000034523453257586, 0.00000000000005000535, 0.00000000345635785295, 0.00000002112341863314, 0.00000024000000872668, 0.00000034523453377314, 0.00000000000004813184, 0.00000000345635686215, 0.00000002112341768074, 0.00000024000000874356, 0.00000034523453389326, 0.00000000000004903036, 0.00000000345635776067, 0.00000002112341768074, 0.00000024000000956720, 0.00000034523453377314, 0.00000000000004989727, 0.00000000345635774487, 0.00000002112341875961, 0.00000024000000874356, 0.00000034523453389326, 0.00000000000005079578, 0.00000000345635864338, 0.00000002112341875961, 0.00000024000000956720]
-    expected_allIntervals_14 = [np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.03945555,  0.53867718], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-1.,         -0.45976608], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.03945555,  0.53867718], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.03945555,  0.46851871], [ 0.71894518,  1.        ]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.41480102,  0.71894518]]), np.array([[ 0.53867718,  1.        ], [-0.45976608,  0.03945555], [ 0.53867718,  1.        ], [ 0.46851871,  0.8650092 ], [ 0.71894518,  1.        ]])]
-    expected_transforms_14 = [[np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[ 0.51972777,  1.,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[ 1.,  1.,  0.51972777,  1.,  1.], [ 0.,  0., -0.48027223,  0.,  0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[ 1.,  1.,  1.,  0.51972777,  1.], [ 0.,  0.,  0., -0.48027223,  0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[ 1.,  1.,  1.,  1.,  0.51972777], [ 0.,  0.,  0.,  0., -0.48027223]])], [np.array([[0.48027223, 1., 1., 1., 1.], [0.51972777, 0., 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]]), np.array([[1., 1., 0.48027223, 1., 1.], [0., 0., 0.51972777, 0., 0.]]), np.array([[1., 1., 1., 0.48027223, 1.], [0., 0., 0., 0.51972777, 0.]]), np.array([[1., 1., 1., 1., 0.48027223], [0., 0., 0., 0., 0.51972777]])]]
-    allMs_14, allErrors_14, allIntervals_14 = CS.getSubdivisionIntervals(Ms_14,errors_14,trackedInterval_14,False,level_14)
-    allMs_first_14 = [M[0,0,0,0,0] for Ms in allMs_14 for M in Ms]
-    allMs_shapes_14 = [M.shape for Ms in allMs_14 for M in Ms]
-    allErrors_flattened_14 = [item for error_arr in allErrors_14 for item in error_arr]
-    allTransforms_14 = [interval.transforms for interval in allIntervals_14]
-    assert np.allclose(expected_allMs_first_14,allMs_first_14)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_14,expected_allMs_shapes_14)])
-    assert np.allclose(expected_allErrors_flattened_14,allErrors_flattened_14)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_14,expected_allIntervals_14))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_14,expected_transforms_14))
-
-    trackedInterval_15 = CS.TrackedInterval(np.array([[ 0.03945555,  0.03945555 + 1e-5        ],
-                                                        [-1.,          1],
-                                                        [ 1.,  1. + 1e-5        ],
-                                                        [ 0.03945555,  0.03945555 + 1e-5 ],
-                                                        [ 0.41480102,  0.41480102 + 1e-5       ]]))
-    Ms_15 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-    level_15 = 2
-    errors_15 = np.array([.00000034523452345,.000000000000035634,.0000000034563456,.00000002112341243,.00000024])
-    expected_allMs_first_15 = [0.84962363920314454280, 1.64532289992132829681, 1.42183189992132819057, 0.47518696292132828196, 0.89590788792132824980, 1.25879568191777035935, 1.78171358048892480852, 1.55822258048892470228, 0.61157764348892462714, 1.03229856848892476151]
-    expected_allMs_shapes_15 = [(3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3), (3, 3, 3, 2, 2), (3, 3, 3, 3, 3), (3, 3, 3, 3, 3), (3, 3, 3, 3, 2), (3, 3, 3, 2, 3)]
-    expected_allErrors_flattened_15 = [0.00000034523452585998, 0.00000000000003866636, 0.00000000345634820868, 0.00000002112341378635, 0.00000024000000205011, 0.00000034523452585998, 0.00000000000003866636, 0.00000000345634820868, 0.00000002112341378635, 0.00000024000000205011]
-    expected_allIntervals_15 = [np.array([[ 0.03945555,  0.03946555], [-1.,          0.03945555], [ 1.,          1.00001   ], [ 0.03945555,  0.03946555], [ 0.41480102,  0.41481102]]),  np.array([[0.03945555, 0.03946555], [0.03945555, 1.        ], [1.,         1.00001   ], [0.03945555, 0.03946555], [0.41480102, 0.41481102]])]    
-    expected_transforms_15 = [[np.array([[ 1.,  0.51972777,  1.,  1.,  1.], [ 0., -0.48027223,  0.,  0.,  0.]])], [np.array([[1., 0.48027223, 1., 1., 1.], [0., 0.51972777, 0., 0., 0.]])]]
-    allMs_15, allErrors_15, allIntervals_15 = CS.getSubdivisionIntervals(Ms_15,errors_15,trackedInterval_15,False,level_15)
-    allMs_first_15 = [M[0,0,0,0,0] for Ms in allMs_15 for M in Ms]
-    allMs_shapes_15 = [M.shape for Ms in allMs_15 for M in Ms]
-    allErrors_flattened_15 = [item for error_arr in allErrors_15 for item in error_arr]
-    allTransforms_15 = [interval.transforms for interval in allIntervals_15]
-    assert np.allclose(expected_allMs_first_15,allMs_first_15)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_15,expected_allMs_shapes_15)])
-    assert np.allclose(expected_allErrors_flattened_15,allErrors_flattened_15)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_15,expected_allIntervals_15))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_15,expected_transforms_15))
-
-    trackedInterval_16 = CS.TrackedInterval(np.array([[1,1 + 1e-5],
-                                                      [1,10],
-                                                      [1,1 + 1e-5],
-                                                      [1,13]]))
-    Ms_16 = [1.0*np.arange(2*3*2*1).reshape(2,3,2,1),1.0*np.arange(2*3*2*1).reshape(2,3,2,1),1.0*np.arange(2*3*2*1).reshape(2,3,2,1),1.0*np.arange(2*3*2*1).reshape(2,3,2,1)]
-    level_16 = 4
-    errors_16 = np.array([.00000034523452345,.000000000000035634,.0000000034563456,.00000002112341243])
-    expected_allMs_first_16 = [-2.03478532688930568995, -2.03478532688930568995, -2.03478532688930568995, -2.03478532688930568995, 0.12303686350311315856, 0.12303686350311315856, 0.12303686350311315856, 0.12303686350311315856]
-    expected_allMs_shapes_16 = [(2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1)]
-    expected_allErrors_flattened_16 = [0.00000034523456741483, 0.00000000000007959883, 0.00000000345638956483, 0.00000002112345639483, 0.00000034523456741483, 0.00000000000007959883, 0.00000000345638956483, 0.00000002112345639483]
-    expected_allIntervals_16 = [np.array([[ 1.,          1.00001   ], [ 1.,          5.67754996], [ 1.,          1.00001   ], [ 1.,         13.        ]]), np.array([[ 1.,          1.00001   ], [ 5.67754996, 10.        ], [ 1.,          1.00001   ], [ 1.,         13.        ]])]
-    expected_transforms_16 = [[np.array([[ 1.,  0.51972777,  1.,  1.], [ 0., -0.48027223,  0.,  0.]])], [np.array([[1., 0.48027223, 1., 1.], [0., 0.51972777, 0., 0.]])]]
-    allMs_16, allErrors_16, allIntervals_16 = CS.getSubdivisionIntervals(Ms_16,errors_16,trackedInterval_16,False,level_16)
-    allMs_first_16 = [M[0,0,0,0] for Ms in allMs_16 for M in Ms]
-    allMs_shapes_16 = [M.shape for Ms in allMs_16 for M in Ms]
-    allErrors_flattened_16 = [item for error_arr in allErrors_16 for item in error_arr]
-    allTransforms_16 = [interval.transforms for interval in allIntervals_16]
-    assert np.allclose(expected_allMs_first_16,allMs_first_16)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_16,expected_allMs_shapes_16)])
-    assert np.allclose(expected_allErrors_flattened_16,allErrors_flattened_16)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_16,expected_allIntervals_16))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_16,expected_transforms_16))
-
-    trackedInterval_17 = CS.TrackedInterval(np.array([[1,5],
-                                                      [1,10],
-                                                      [1,1 + 1e-5],
-                                                      [1,13]]))
-    Ms_17 = [1.0*np.arange(2*3*2*1).reshape(2,3,2,1),1.0*np.arange(2*3*2*1).reshape(2,3,2,1),1.0*np.arange(2*3*2*1).reshape(2,3,2,1),1.0*np.arange(2*3*2*1).reshape(2,3,2,1)]
-    level_17 = 4
-    errors_17 = np.array([.00000034523452345,.000000000000035634,.0000000034563456,.00000002112341243])
-    expected_allMs_first_17 = [-1.78130725334957440431, -1.78130725334957440431, -1.78130725334957440431, -1.78130725334957440431, -3.65515690937392889737, -3.65515690937392889737, -3.65515690937392889737, -3.65515690937392889737, -2.30908724917568131119, -2.30908724917568131119, -2.30908724917568131119, -2.30908724917568131119, 4.21161857078101142804, 4.21161857078101142804, 4.21161857078101142804, 4.21161857078101142804]
-    expected_allMs_shapes_17 = [(2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1), (2, 3, 2, 1)]
-    expected_allErrors_flattened_17 = [0.00000034523458129066, 0.00000000000009347466, 0.00000000345640344066, 0.00000002112347027066, 0.00000034523459672472, 0.00000000000010890872, 0.00000000345641887472, 0.00000002112348570472, 0.00000034523458129066, 0.00000000000009347466, 0.00000000345640344066, 0.00000002112347027066, 0.00000034523459672472, 0.00000000000010890872, 0.00000000345641887472, 0.00000002112348570472]
-    expected_allIntervals_17 = [np.array([[ 1.,          3.0789111 ], [ 1.,          5.67754996], [ 1.,          1.00001   ], [ 1.,         13.        ]]), np.array([[ 1.,          3.0789111 ], [ 5.67754996, 10.        ], [ 1.,          1.00001   ], [ 1.,         13.        ]]), np.array([[ 3.0789111,   5.        ], [ 1.,          5.67754996], [ 1.,          1.00001   ], [ 1.,         13.        ]]), np.array([[ 3.0789111,   5.        ], [ 5.67754996, 10.        ], [ 1.,          1.00001   ], [ 1.,         13.        ]])]
-    expected_transforms_17 = [[np.array([[ 0.51972777,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.]]), np.array([[ 1.,  0.51972777,  1.,  1.], [ 0., -0.48027223,  0.,  0.]])], [np.array([[ 0.51972777,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.]]), np.array([[1., 0.48027223, 1., 1.], [0., 0.51972777, 0., 0.]])], [np.array([[0.48027223, 1., 1., 1.], [0.51972777, 0., 0., 0.]]), np.array([[ 1.,  0.51972777,  1.,  1.], [ 0., -0.48027223,  0.,  0.]])], [np.array([[0.48027223, 1., 1., 1.], [0.51972777, 0., 0., 0.]]), np.array([[1., 0.48027223, 1., 1.], [0., 0.51972777, 0., 0.]])]]
-    allMs_17, allErrors_17, allIntervals_17 = CS.getSubdivisionIntervals(Ms_17,errors_17,trackedInterval_17,False,level_17)
-    allMs_first_17 = [M[0,0,0,0] for Ms in allMs_17 for M in Ms]
-    allMs_shapes_17 = [M.shape for Ms in allMs_17 for M in Ms]
-    allErrors_flattened_17 = [item for error_arr in allErrors_17 for item in error_arr]
-    allTransforms_17 = [interval.transforms for interval in allIntervals_17]
-    assert np.allclose(expected_allMs_first_17,allMs_first_17)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_17,expected_allMs_shapes_17)])
-    assert np.allclose(expected_allErrors_flattened_17,allErrors_flattened_17)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_17,expected_allIntervals_17))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_17,expected_transforms_17))
-
-    trackedInterval_18 = CS.TrackedInterval(np.array([[1,5],
-                                                      [1,10],
-                                                      [1,1 + 1e-5],
-                                                      [1,13]]))
-    Ms_18 = [1.0*np.arange(15*3*2*1).reshape(15,2,3,1),1.0*np.arange(15*1*4*3).reshape(15,1,4,3),1.0*np.arange(15*3*2*1).reshape(15,3,2,1),1.0*np.arange(15*2*2*1).reshape(15,2,2,1)]
-    level_18 = 4
-    errors_18 = np.array([.00000034523452345,.000000000000035634,.0000000034563456,.00000002112341243])
-    expected_allMs_first_18 = [-10.01235914199165932814, -20.02471828398331865628, -10.01235914199165932814, -6.67490609466110651482, 5.13004204674147068488, 10.26008409348294136976, 5.13004204674147068488, 3.42002803116097897629]
-    expected_allMs_shapes_18 = [(15, 2, 3, 1), (15, 1, 4, 3), (15, 3, 2, 1), (15, 2, 2, 1), (15, 2, 3, 1), (15, 1, 4, 3), (15, 3, 2, 1), (15, 2, 2, 1)]
-    expected_allErrors_flattened_18 = [0.00000034524786277964, 0.00000000005369271278, 0.00000000346968492964, 0.00000002112930771426, 0.00000034524786277964, 0.00000000005369271278, 0.00000000346968492964, 0.00000002112930771426]    
-    expected_allIntervals_18 = [np.array([[ 1.,         3.0789111], [ 1.,        10.       ], [ 1.,         1.00001  ], [ 1.,        13.       ]]), np.array([[ 3.0789111,  5.       ],      [ 1.,        10.       ], [ 1.,         1.00001  ], [ 1.,        13.       ]])]
-    expected_transforms_18 = [[np.array([[ 0.51972777,  1.,  1.,  1.], [-0.48027223,  0.,  0.,  0.]])], [np.array([[0.48027223, 1., 1., 1.], [0.51972777, 0., 0., 0.]])]]
-    allMs_18, allErrors_18, allIntervals_18 = CS.getSubdivisionIntervals(Ms_18,errors_18,trackedInterval_18,False,level_18)
-    allMs_first_18 = [M[0,0,0,0] for Ms in allMs_18 for M in Ms]
-    allMs_shapes_18 = [M.shape for Ms in allMs_18 for M in Ms]
-    allErrors_flattened_18 = [item for error_arr in allErrors_18 for item in error_arr]
-    allTransforms_18 = [interval.transforms for interval in allIntervals_18]
-    assert np.allclose(expected_allMs_first_18,allMs_first_18)
-    assert np.all([shape == exp for (shape,exp) in zip(allMs_shapes_18,expected_allMs_shapes_18)])
-    assert np.allclose(expected_allErrors_flattened_18,allErrors_flattened_18)
-    assert np.all(np.allclose(interval.interval,exp) for interval,exp in zip(allIntervals_18,expected_allIntervals_18))
-    assert np.all(np.allclose(transforms,exp) for (transforms,exp) in zip(allTransforms_18,expected_transforms_18))
-
