--- yroots/test_trimMs.py
@@ -1,673 +1,0 @@
-import numpy as np
-# Ms_1 doesn't trim
-Ms_1 = [np.array([[[-1.19628962e-01,  4.78500000e+00,  1.36125000e+00],       
-                        [-1.22216380e+00, -5.88038994e-17, -1.29732624e-16]],       
-
-                    [[ 2.75448343e+01, -4.09625873e-15, -5.24707922e-16],        
-                        [ 1.38346959e+01, -1.64018392e-15, -9.14780575e-16]],       
-
-                    [[-2.34899407e+00, -5.34145567e-16, -3.02474728e-16],        
-                        [-1.00260753e-15,  1.73733322e-16,  3.97924039e-17]],       
-
-                    [[ 4.43170528e+00,  1.12302734e-16,  4.10866620e-16],        
-                        [ 9.31305763e-16, -1.86261153e-16, -1.86261153e-16]]]), 
-
-            np.array([[[ 2.84845613e+02,  1.65000000e+00],
-                        [ 2.27137940e+02, -1.72605562e-13],
-                        [ 1.29793108e+02,  2.54414728e-14],
-                        [ 4.86724156e+01,  6.09220633e-15],
-                        [ 1.08160924e+01,  1.01171420e-14],
-                        [ 1.08160924e+00, -3.13565083e-15]],
-
-                    [[-9.64091464e+01, -2.86049313e-14],
-                        [ 4.14691757e-14,  7.88666604e-14],
-                        [ 2.67503292e-14, -9.27142549e-15],
-                        [-6.34389840e-14, -6.91944204e-15],
-                        [-1.24333103e-14, -2.02584903e-15],
-                        [ 6.92587964e-15,  8.01648088e-16]],
-
-                    [[ 1.93541862e+02, -5.17408223e-14],
-                        [-5.39535253e-14, -1.96574197e-14],
-                        [-5.00488782e-14, -4.71474082e-14],
-                        [ 2.06908008e-14,  8.68344641e-15],
-                        [ 1.40966336e-14,  2.75871201e-14],
-                        [-3.57269884e-15, -6.35581641e-15]],
-
-                    [[-3.26656222e+01,  1.37816904e-14],
-                        [-1.57824001e-14, -1.92527602e-14],
-                        [ 1.07754349e-15,  7.17861036e-15],
-                        [ 2.10137810e-14, -6.69939041e-16],
-                        [-4.24524051e-15, -2.25717150e-16],
-                        [-3.59823275e-16,  3.49127175e-17]],
-
-                    [[ 4.62211928e+01,  4.21235493e-15],
-                        [-7.00784374e-16,  8.21968518e-15],
-                        [ 4.49894444e-15,  1.37359712e-14],
-                        [ 1.55416701e-15, -5.19175671e-15],
-                        [-2.74045293e-15, -6.88588808e-15],
-                        [ 4.70817493e-16,  1.72633081e-15]]]), 
-
-            np.array([[[ 4.68071678e+00, -4.78500000e+00, -1.36125000e+00],
-                        [ 6.25613546e+00, -2.64267077e-17, -1.19440696e-16],        
-                        [ 1.56403386e+00,  7.59312342e-18,  4.78427451e-17]],       
-
-                    [[-1.20118630e+00, -1.72239045e-16, -1.51455377e-16],        
-                        [ 5.52210332e-16, -3.13691772e-16, -1.22891472e-16],        
-                        [-8.91891458e-17,  9.01323150e-17,  3.09872741e-17]],       
-
-                    [[ 3.39930849e+00,  0.00000000e+00,  3.51928144e-17],        
-                        [ 1.00761893e-16,  0.00000000e+00,  1.00761893e-16],        
-                        [-4.17799856e-17,  0.00000000e+00, -4.17799856e-17]]])]
-errors_1 = np.array([.000000132,.000000000050623,.00000000000203045])
-# Ms_2 doesn't trim
-Ms_2 = [np.array([[[[[ 1.07142165e+00,  1.61772562e-01],
-          [ 5.75461946e-02,  5.61619043e-01]],
-
-         [[ 2.96651313e-01, -3.39708980e-17],
-          [ 6.23974136e-17,  3.06138888e-17]],
-
-         [[ 3.35269245e-01, -2.14750436e-18],
-          [ 1.63404061e-17,  3.92462018e-18]]],
-
-
-        [[[ 4.05175438e-01,  1.49089271e-17],
-          [-2.42704359e-17,  6.11752001e-18]],
-
-         [[ 8.91612496e-17, -1.64457404e-17],
-          [ 4.37735368e-18, -8.26735644e-18]],
-
-         [[ 1.59746311e-17, -5.58287457e-18],
-          [ 1.21413250e-18, -2.15690973e-18]]],
-
-
-        [[[ 1.01293860e-01, -2.91078151e-18],
-          [ 6.02650904e-18, -7.69091844e-19]],
-
-         [[-1.61151471e-17,  1.96422369e-18],
-          [-1.02880386e-18,  1.16696315e-18]],
-
-         [[-3.16461064e-18,  1.07109611e-18],
-          [-3.19107512e-19,  4.03906181e-19]]]],
-
-
-
-       [[[[ 4.99963263e-01,  7.24023441e-18],
-          [ 2.08333217e-17, -3.58422625e-18]],
-
-         [[-2.88911748e-17, -2.39516560e-17],
-          [-3.52016062e-17,  2.07163246e-17]],
-
-         [[-2.99141286e-17, -9.60551746e-18],
-          [-1.09122488e-17,  5.52770123e-18]]],
-
-
-        [[[ 6.41527705e-17, -4.54209112e-18],
-          [ 1.38295961e-17, -2.04590576e-18]],
-
-         [[-3.54019101e-17, -6.07501389e-18],
-          [-6.98105270e-18, -1.19866640e-18]],
-
-         [[-1.31610394e-17, -1.04908960e-18],
-          [-2.96892590e-18, -1.65904797e-19]]],
-
-
-        [[[-1.29567793e-17, -6.01707661e-19],
-          [-4.40294636e-18, -2.94442193e-19]],
-
-         [[ 2.92589264e-17,  3.61843830e-18],
-          [ 5.62920573e-18,  5.67975389e-19]],
-
-         [[ 5.75914328e-18,  7.63941973e-19],
-          [ 1.41156695e-18,  1.78514600e-19]]]],
-
-
-
-       [[[[ 1.27152103e-01,  1.91152932e-18],
-          [ 4.79121097e-18, -1.23937605e-18]],
-
-         [[-2.09329083e-17, -4.09015035e-18],
-          [-7.01631281e-18,  3.73102191e-18]],
-
-         [[-9.06126759e-18, -1.92681844e-18],
-          [-2.12745553e-18,  1.25546687e-18]]],
-
-
-        [[[ 1.36153101e-17, -1.23234013e-18],
-          [ 2.89822838e-18, -5.16802841e-19]],
-
-         [[-7.90847046e-18, -1.37345331e-18],
-          [-1.56775419e-18, -2.73832184e-19]],
-
-         [[-2.10526282e-18, -8.93431017e-20],
-          [-4.34200889e-19,  2.74305342e-21]]],
-
-
-        [[[-3.91222649e-18, -1.41462319e-19],
-          [-9.76066997e-19, -1.69143790e-20]],
-
-         [[ 7.54787397e-18,  8.63779933e-19],
-          [ 1.31538564e-18,  1.03280515e-19]],
-
-         [[ 1.46097501e-18,  1.51097345e-19],
-          [ 2.70074762e-19,  1.80664207e-20]]]]]), 
-          
-          np.array([[[[[ 1.71925557e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 4.41708911e-01,  3.45276625e-17,  2.71584393e-18],
-          [ 3.83365591e-01,  5.51328328e-18,  2.19183835e-19]],
-
-         [[ 2.96651313e-01,  5.68098934e-17,  1.04881815e-17],
-          [-2.59134440e-17,  5.02067534e-17,  3.18884982e-18],
-          [-4.40610755e-17,  1.85375284e-17,  4.80625344e-18]],
-
-         [[ 3.35269245e-01, -3.38256162e-18, -2.81267507e-18],
-          [-3.88863843e-18,  1.97375581e-18, -5.95518601e-20],
-          [-2.82334822e-18,  1.92548058e-18,  7.35266273e-19]]],
-
-
-        [[[ 1.35058479e-01, -9.34504427e-18, -9.29019885e-19],
-          [-1.10973940e-17,  1.03984213e-17,  1.13091038e-18],
-          [-2.27373541e-18,  1.44719312e-18,  1.66375006e-19]],
-
-         [[-3.86160966e-17,  3.43535302e-17,  3.62552216e-18],
-          [-2.65620582e-17,  2.40509961e-17,  2.55367163e-18],
-          [-2.24963724e-18,  2.14897426e-18,  2.27589614e-19]],
-
-         [[-7.29423233e-18,  5.54979427e-18,  5.98695706e-19],
-          [-3.99966511e-18,  3.36022405e-18,  3.59624925e-19],
-          [ 7.16837006e-20,  4.65715043e-19,  4.02772780e-20]]],
-
-
-        [[[ 3.37646199e-02,  6.57761381e-18,  1.05638586e-18],
-          [-1.54799727e-18,  1.72457255e-18,  5.76470040e-19],
-          [ 1.18536415e-19,  9.64716050e-21,  4.27921252e-20]],
-
-         [[-4.17762847e-18,  6.66283925e-19,  5.32216329e-19],
-          [ 6.92434840e-20,  6.53808761e-20,  4.41203718e-19],
-          [ 1.78340075e-19, -2.27569503e-20,  4.26724892e-20]],
-
-         [[-8.56908140e-20, -9.72518670e-20,  4.13326424e-20],
-          [ 2.87748879e-20, -1.15988381e-19,  4.08861804e-20],
-          [-3.94480233e-20, -3.98168367e-20,  2.69575387e-21]]]],
-
-
-
-       [[[[ 4.99963263e-01,  2.29272093e-17, -2.86799058e-19],
-          [ 2.14831673e-18,  3.13193909e-17,  3.24546778e-18],
-          [ 1.35741026e-17, -2.13537735e-19, -6.57646260e-19]],
-
-         [[ 2.78293091e-16, -1.38063712e-16, -2.50498941e-17],
-          [ 6.04556822e-17,  4.57639616e-17,  2.46885166e-18],
-          [-8.79903230e-18,  1.70212431e-17,  2.95411077e-18]],
-
-         [[ 1.50741235e-17, -1.98538344e-17, -3.89293146e-18],
-          [ 5.49952480e-18,  2.84087843e-18,  5.07447260e-20],
-          [-2.09725092e-18,  2.74694437e-18,  5.20118981e-19]]],
-
-
-        [[[-3.36484194e-17,  3.10696010e-17,  3.20162611e-18],
-          [ 3.56944474e-17, -3.25664835e-17, -3.36584778e-18],
-          [ 6.31873340e-18, -6.34800243e-18, -6.46404749e-19]],
-
-         [[ 2.81977474e-17, -2.38792491e-17, -2.49869943e-18],
-          [-2.90390317e-17,  2.59350981e-17,  2.68616991e-18],
-          [-5.43349495e-18,  4.87683452e-18,  5.04862589e-19]],
-
-         [[ 5.77179529e-18, -5.99247884e-18, -6.07358261e-19],
-          [-6.79088336e-18,  5.97713108e-18,  6.20757732e-19],
-          [-9.18819151e-19,  1.32384739e-18,  1.28793051e-19]]],
-
-
-        [[[ 1.40974138e-21, -3.54630380e-18, -3.19792116e-19],
-          [-5.73760846e-18,  1.99122869e-18,  2.46616928e-19],
-          [-7.98893529e-19,  4.72947818e-19,  5.17284702e-20]],
-
-         [[ 2.34561328e-18,  9.94655571e-18,  8.17713030e-19],
-          [ 3.56484031e-18, -5.49916153e-18, -5.44390129e-19],
-          [ 4.22030873e-19, -1.23719330e-18, -1.15292452e-19]],
-
-         [[-2.32025897e-19,  1.48923258e-18,  1.31516094e-19],
-          [ 1.28054485e-18, -8.01145834e-19, -9.02892285e-20],
-          [ 1.55691965e-19, -2.06043501e-19, -2.04047658e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  7.47509277e-19, -1.68244688e-18],
-          [ 4.63304008e-19,  6.35723598e-18,  6.79396127e-19],
-          [ 4.94612170e-18,  1.46994270e-19, -3.79295338e-20]],
-
-         [[ 5.91928627e-17, -2.87891070e-17, -3.32217995e-18],
-          [ 1.99411391e-17,  4.96747918e-18,  2.02129771e-19],
-          [ 4.24216090e-19,  1.98140464e-18,  1.73282230e-19]],
-
-         [[ 5.15689754e-18, -3.87702672e-18, -4.06527421e-19],
-          [ 2.23891020e-18, -2.45004990e-19, -4.67452746e-20],
-          [-1.12657737e-19,  3.32790994e-19,  3.14783904e-20]]],
-
-
-        [[[-1.04290640e-17,  9.34172522e-18,  9.57292638e-19],
-          [ 5.45480277e-18, -5.04506206e-18, -5.31161827e-19],
-          [ 1.24687631e-18, -1.14042138e-18, -1.18963104e-19]],
-
-         [[ 8.80257271e-18, -7.54482795e-18, -7.98676635e-19],
-          [-4.43106195e-18,  3.94672772e-18,  3.97201471e-19],
-          [-1.02679171e-18,  9.02116107e-19,  9.24464753e-20]],
-
-         [[ 1.98766428e-18, -1.75969853e-18, -1.83989290e-19],
-          [-1.01078505e-18,  9.28863753e-19,  9.44209950e-20],
-          [-2.33369611e-19,  2.11640682e-19,  2.17289287e-20]]],
-
-
-        [[[ 1.22499456e-18, -1.83448985e-18, -2.47020324e-19],
-          [-8.20959421e-19,  7.57803612e-20, -5.45301161e-20],
-          [-1.73151346e-19,  8.84469260e-20,  2.32250876e-21]],
-
-         [[ 1.18291130e-18,  1.77850549e-18,  6.21181689e-20],
-          [ 6.67279506e-19, -1.05968248e-18, -1.81474744e-19],
-          [ 4.90292034e-20, -2.31806698e-19, -2.94914098e-20]],
-
-         [[-2.91321091e-20,  3.00984336e-19,  1.77874793e-20],
-          [ 2.39009743e-19, -1.32272646e-19, -2.44110439e-20],
-          [ 3.66238103e-20, -3.22597475e-20, -4.36406276e-21]]]]]), 
-          np.array([[[[[ 1.49576457e+00,  1.68606883e-01,  4.11377490e-01],
-          [ 2.20854456e-01, -5.06979570e-17, -1.48297754e-17],
-          [ 1.91682796e-01, -4.28098704e-18, -3.03250631e-18]],
-
-         [[ 2.96651313e-01, -1.52117951e-16, -2.87974276e-17],
-          [ 9.85544813e-18, -1.26068312e-17, -1.82987398e-18],
-          [-2.64947010e-17,  2.04151968e-17,  3.43762545e-18]],
-
-         [[ 3.35269245e-01, -2.66571677e-17, -8.65114674e-18],
-          [ 1.41002534e-18, -1.22337404e-18, -2.49667270e-19],
-          [-2.97331670e-18,  2.22719784e-18,  5.02129829e-19]]],
-
-
-        [[[ 1.35058479e-01,  6.33671440e-17,  5.31519365e-18],
-          [-2.74298794e-18,  4.29406209e-18,  1.07929091e-19],
-          [ 7.93333491e-18, -7.50055855e-18, -1.08630334e-19]],
-
-         [[-6.00791689e-17,  5.54051086e-17,  6.73924499e-18],
-          [-3.34756515e-18,  1.59372967e-18,  4.37359936e-19],
-          [ 4.77460128e-18, -4.13673986e-18, -9.68554646e-19]],
-
-         [[-5.74073182e-18,  5.44263054e-18,  7.95821283e-19],
-          [-2.10323124e-19, -1.43941116e-19,  4.81126567e-20],
-          [-5.98371220e-20,  1.02228055e-19, -1.14428425e-19]]],
-
-
-        [[[ 3.37646199e-02, -2.37467886e-18,  5.05147144e-19],
-          [ 8.46931293e-19, -6.83729888e-20,  1.59692136e-19],
-          [-3.52574395e-18,  3.41684249e-18, -3.54198703e-20]],
-
-         [[ 7.09803562e-19, -1.78247073e-18, -7.62390099e-19],
-          [-9.61353501e-19,  4.92268856e-19, -8.76704803e-20],
-          [-1.98866700e-18,  1.21408536e-18,  4.48047751e-19]],
-
-         [[-4.04403498e-19,  2.28264502e-19, -1.11996935e-19],
-          [-3.11471188e-19,  1.82498799e-19, -1.29409342e-20],
-          [ 8.58626558e-20, -1.87289179e-19,  5.47037359e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01,  8.32050447e-17, -6.85647337e-19],
-          [ 5.53515005e-17,  2.27049024e-17,  1.05680764e-18],
-          [ 1.99603570e-18,  1.63225390e-17,  1.59012394e-18]],
-
-         [[ 8.83607766e-17, -9.99996893e-17, -1.16883354e-17],
-          [ 1.63971286e-17,  3.14925071e-18, -2.55450992e-19],
-          [ 1.18374815e-18,  1.02932357e-17,  1.49308097e-18]],
-
-         [[-1.32580578e-18, -3.25472901e-17, -6.36865089e-18],
-          [ 1.00480865e-18, -6.17269549e-19, -1.92308735e-19],
-          [-8.47901575e-19,  1.65871216e-18,  4.24018265e-19]]],
-
-
-        [[[ 1.02499832e-17,  2.91175683e-18,  8.69254265e-20],
-          [-4.64903769e-18, -2.33254419e-18, -1.36843898e-19],
-          [-1.00984279e-18, -6.12639718e-19, -3.11328996e-20]],
-
-         [[-5.58011466e-18, -2.77533450e-18, -1.47196267e-19],
-          [ 3.45566799e-18,  2.25499877e-18,  1.47268742e-19],
-          [ 7.49863393e-19,  4.11677340e-19,  1.88624856e-20]],
-
-         [[-1.48455723e-18, -5.98554534e-19, -2.73877700e-20],
-          [ 8.14791283e-19,  5.01547273e-19,  3.20893871e-20],
-          [ 1.87092116e-19,  9.22429115e-20,  3.90211539e-21]]],
-
-
-        [[[ 1.57231038e-17, -1.00800814e-17, -1.20004923e-18],
-          [-3.40022589e-18,  1.55870165e-20,  3.04003309e-20],
-          [-4.39902043e-19, -5.25845976e-19,  7.55148791e-21]],
-
-         [[ 5.31534932e-18,  5.48792490e-18,  4.70440868e-19],
-          [ 6.48576503e-19,  2.35966586e-18,  2.13106544e-19],
-          [-1.06759231e-19, -1.39192708e-19, -4.84695289e-20]],
-
-         [[-5.92137415e-20,  1.54844077e-18,  1.52267313e-19],
-          [ 2.58239351e-19,  4.17326986e-19,  3.65022534e-20],
-          [ 6.66077987e-20, -5.10143140e-20, -1.41976204e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  2.64053263e-17,  1.87453686e-18],
-          [ 1.14621989e-17,  6.36289187e-18,  4.76910862e-19],
-          [ 2.80365728e-18,  1.82800781e-18, -8.67438264e-20]],
-
-         [[ 2.07717264e-17, -2.28941320e-17, -2.68751498e-18],
-          [ 3.56117578e-18,  9.58256645e-19, -3.55695508e-20],
-          [ 2.00680115e-19,  2.38975911e-18,  3.47044517e-19]],
-
-         [[ 2.85446168e-18, -7.75745683e-18, -1.53912412e-18],
-          [ 2.30510206e-19, -1.38180388e-19, -4.40186598e-20],
-          [-2.58668313e-19,  4.33721110e-19,  1.03286499e-19]]],
-
-
-        [[[ 2.39508763e-18,  6.30999482e-19,  1.60529540e-20],
-          [-1.03579320e-18, -5.77663838e-19, -3.55574857e-20],
-          [-2.22755939e-19, -1.42123716e-19, -7.40081569e-21]],
-
-         [[-1.25977109e-18, -6.86260230e-19, -3.83648534e-20],
-          [ 8.39703446e-19,  4.78969772e-19,  2.97547248e-20],
-          [ 1.84638527e-19,  9.44665604e-20,  4.13205533e-21]],
-
-         [[-3.39507811e-19, -1.43588952e-19, -6.84045187e-21],
-          [ 1.93105222e-19,  1.10924750e-19,  6.91085159e-21],
-          [ 4.45650185e-20,  2.12347479e-20,  8.74910315e-22]]],
-
-
-        [[[ 3.81102471e-18, -2.60321014e-18, -3.03770305e-19],
-          [-5.33716027e-19, -2.51925780e-19, -1.89508235e-20],
-          [-3.22141721e-20, -1.25374252e-19,  3.71580813e-22]],
-
-         [[ 1.42128275e-18,  9.70789600e-19,  7.99695564e-20],
-          [ 4.22316950e-19,  2.68626044e-19,  2.11255965e-20],
-          [ 5.04478629e-20, -3.63688189e-20, -1.26831797e-20]],
-
-         [[ 9.08348984e-21,  3.22747781e-19,  3.17962323e-20],
-          [ 9.17421902e-20,  6.38540109e-20,  5.12338319e-21],
-          [ 2.42318218e-20, -1.22813672e-20, -3.45488302e-21]]]]]), 
-          np.array([[[[[ 5.49119633e-01,  2.02719397e-01],
-          [ 2.05265617e-02,  2.00327894e-01],
-          [ 1.78153016e-02,  1.73867494e-01]],
-
-         [[ 3.55981576e-02, -4.79403779e-18],
-          [-3.87882958e-17,  6.12918957e-18],
-          [-1.81021848e-17,  1.13416102e-17]],
-
-         [[ 4.02323093e-02,  1.11973965e-19],
-          [-8.98375531e-18,  1.52152949e-18],
-          [-4.01756286e-18,  1.91735809e-18]]],
-
-
-        [[[ 1.35058479e-01,  5.94760094e-18],
-          [ 7.69395246e-18,  7.06389841e-19],
-          [ 2.51265164e-19,  3.19281748e-20]],
-
-         [[-3.16361745e-17, -5.40901655e-18],
-          [ 3.85207648e-18,  1.95744513e-18],
-          [ 1.90223606e-18,  5.17594123e-19]],
-
-         [[-7.28196516e-18, -1.05789185e-18],
-          [ 8.75899333e-19,  3.51447779e-19],
-          [ 4.36259329e-19,  9.65821278e-20]]],
-
-
-        [[[ 3.37646199e-02, -2.65041304e-18],
-          [-1.89010865e-18, -3.06353560e-19],
-          [ 4.62341252e-19,  6.61993892e-20]],
-
-         [[ 1.47680634e-17,  2.64094224e-18],
-          [-9.90355019e-20, -2.47678807e-19],
-          [-6.36336582e-19, -1.47853220e-19]],
-
-         [[ 3.47046958e-18,  5.69551351e-19],
-          [-5.37519332e-20, -3.58990090e-20],
-          [-1.54051928e-19, -2.92922262e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -4.51000037e-18],
-          [-5.86654108e-17, -3.11421864e-17],
-          [-6.34906134e-18, -8.76327862e-18]],
-
-         [[ 4.93397042e-17,  1.27312132e-17],
-          [ 5.13713052e-17,  1.34337514e-17],
-          [ 1.21566217e-17, -6.04640724e-19]],
-
-         [[ 1.04424704e-17,  2.18900658e-18],
-          [ 1.15803135e-17,  3.55019066e-18],
-          [ 2.62535282e-18,  6.36963537e-19]]],
-
-
-        [[[-8.91062611e-18,  2.88584674e-18],
-          [-5.72097891e-18,  3.69523315e-18],
-          [-4.72182756e-19,  4.25784901e-19]],
-
-         [[ 1.49494096e-17, -1.02282534e-18],
-          [-1.02258364e-17, -5.55059951e-18],
-          [-2.14375543e-18, -7.78990578e-19]],
-
-         [[ 1.64723141e-18, -5.45858823e-19],
-          [-1.41291123e-18, -1.11959301e-18],
-          [-2.78594027e-19, -1.42834149e-19]]],
-
-
-        [[[ 9.85390930e-19, -4.16044371e-19],
-          [ 7.78935935e-18,  7.15131094e-19],
-          [ 1.11212800e-18,  1.23424928e-19]],
-
-         [[-4.83095371e-18, -3.52743064e-19],
-          [ 1.89623903e-18,  8.28398167e-19],
-          [ 4.84195607e-19,  1.37535144e-19]],
-
-         [[-4.91299177e-19,  2.55736473e-20],
-          [ 1.09690662e-19,  1.35102418e-19],
-          [ 3.69267458e-20,  1.89322089e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01, -2.14404266e-18],
-          [-9.60099826e-18, -6.27568893e-18],
-          [-3.18297026e-19, -1.19267467e-18]],
-
-         [[ 7.40356148e-18,  4.37533456e-18],
-          [ 8.23054172e-18,  2.41916808e-18],
-          [ 2.14311438e-18, -1.06172510e-18]],
-
-         [[ 1.50386501e-18,  8.39486389e-19],
-          [ 1.97314415e-18,  6.85752285e-19],
-          [ 3.74430051e-19, -7.92905809e-20]]],
-
-
-        [[[-2.31404570e-18,  6.22905375e-19],
-          [-1.18736968e-18,  8.76691280e-19],
-          [-7.84391585e-20,  1.03616949e-19]],
-
-         [[ 3.48968459e-18, -2.30115176e-19],
-          [-2.38133525e-18, -1.28555312e-18],
-          [-4.99577208e-19, -1.80704330e-19]],
-
-         [[ 5.99992360e-19, -8.84979160e-20],
-          [-4.41757350e-19, -2.78416400e-19],
-          [-9.06815078e-20, -3.75082611e-20]]],
-
-
-        [[[ 2.61512218e-19, -9.03113175e-20],
-          [ 1.78170536e-18,  1.62159821e-19],
-          [ 2.52863084e-19,  2.78176640e-20]],
-
-         [[-1.12083114e-18, -8.23100813e-20],
-          [ 4.40583130e-19,  1.91807899e-19],
-          [ 1.12432587e-19,  3.18717451e-20]],
-
-         [[-1.42587480e-19,  9.23368410e-22],
-          [ 4.06294772e-20,  3.38315346e-20],
-          [ 1.20195434e-20,  4.97159986e-21]]]]]), 
-          np.array([[[[[ 9.69840558e-01,  1.68606883e-01,  4.11377490e-01],
-          [ 2.27901995e-01, -1.24309595e-17,  8.25603419e-18]],
-
-         [[ 5.29073342e-02, -7.46091880e-17, -1.85281254e-17],
-          [ 1.83676181e-01, -1.04166859e-16, -1.14069672e-17]],
-
-         [[ 5.97947866e-02, -6.64017247e-18, -6.41927440e-18],
-          [ 2.07587061e-01, -2.54512810e-17, -3.40085408e-18]]],
-
-
-        [[[ 1.35058479e-01,  2.31430691e-18, -6.72762893e-18],
-          [ 1.98802250e-17, -3.18309247e-18, -2.29197550e-18]],
-
-         [[-5.12474635e-17,  6.01397046e-18,  6.20055503e-18],
-          [-1.53997395e-17,  1.58437355e-18,  1.76388565e-18]],
-
-         [[-1.33638081e-17, -1.70368117e-18,  1.18667316e-18],
-          [-5.46909963e-18,  9.67832915e-19,  4.87959566e-19]]],
-
-
-        [[[ 3.37646199e-02, -6.81868016e-18,  2.73831818e-19],
-          [-5.93822658e-18, -1.56759129e-18,  1.24806388e-19]],
-
-         [[ 1.33845025e-17,  4.53207270e-18, -2.90078414e-19],
-          [ 4.24065660e-18,  1.06109094e-18, -1.09252487e-19]],
-
-         [[ 3.30419645e-18,  1.99052380e-18,  1.49874169e-20],
-          [ 1.31709634e-18, -1.84937423e-19, -7.00129850e-20]]]],
-
-
-
-       [[[[ 4.99963263e-01, -1.00662370e-17,  2.29723215e-18],
-          [-4.39869872e-17,  3.82903378e-17,  7.46065937e-18]],
-
-         [[ 1.77788932e-16, -1.39729683e-16, -2.46306662e-17],
-          [ 7.94939073e-17, -6.65767506e-17, -9.80600214e-18]],
-
-         [[ 1.16756624e-17, -2.06386167e-17, -4.02738435e-18],
-          [ 6.48364192e-18, -1.24755483e-17, -1.99568033e-18]]],
-
-
-        [[[ 4.59654126e-18,  1.97245530e-18,  4.27583244e-20],
-          [ 1.13417236e-18,  6.43430431e-19,  1.93352139e-20]],
-
-         [[-3.43457551e-19, -3.49061114e-20,  5.27146302e-20],
-          [-2.50400329e-19,  1.18720804e-19,  2.88122201e-20]],
-
-         [[-2.84298440e-18, -1.55926603e-18, -8.35910277e-20],
-          [-8.75261679e-19, -4.42063177e-19, -2.04385641e-20]]],
-
-
-        [[[ 8.01750108e-18,  4.16761857e-19, -4.80192786e-19],
-          [ 1.14385235e-18,  6.63717382e-19, -8.91873113e-20]],
-
-         [[ 8.78699897e-19,  1.54456738e-18,  4.54696849e-19],
-          [-7.56470482e-19,  1.28629194e-18,  2.20321242e-19]],
-
-         [[ 3.91057974e-19,  5.30136140e-19,  1.03753213e-19],
-          [-3.31330633e-19,  1.58168913e-19,  5.13634639e-20]]]],
-
-
-
-       [[[[ 1.27152103e-01,  6.91300878e-18,  2.11366792e-18],
-          [-9.64263403e-18,  9.23999004e-18,  1.77144208e-18]],
-
-         [[ 3.59148864e-17, -3.59563826e-17, -5.99397999e-18],
-          [ 1.67605442e-17, -1.64040837e-17, -2.34782127e-18]],
-
-         [[ 3.90764224e-18, -6.15476863e-18, -1.08313730e-18],
-          [ 1.00134007e-18, -3.14625746e-18, -4.80619661e-19]]],
-
-
-        [[[ 2.97186772e-19, -8.53943022e-20, -3.92173999e-20],
-          [ 4.77495278e-20,  9.55773189e-21, -8.25519959e-21]],
-
-         [[ 2.31408072e-20,  1.30026992e-19,  3.35763546e-20],
-          [-2.59855819e-20,  6.63921673e-20,  1.26728393e-20]],
-
-         [[-1.61954334e-20,  3.05380252e-20,  7.64563700e-21],
-          [-3.49797155e-21,  7.20620189e-21,  2.54161744e-21]]],
-
-
-        [[[ 1.95734477e-18, -5.62427761e-19, -2.58295388e-19],
-          [ 3.14490069e-19,  6.29495599e-20, -5.43707636e-20]],
-
-         [[ 1.52411016e-19,  8.56389571e-19,  2.21142084e-19],
-          [-1.71147398e-19,  4.37275053e-19,  8.34664190e-20]],
-
-         [[-1.06667086e-19,  2.01130903e-19,  5.03560353e-20],
-          [-2.30384962e-20,  4.74618082e-20,  1.67397141e-20]]]]])]
-errors_2 = np.array([4500000000000000,623000000000000000000,55400000000000000,4500000000000000,4500000000000000])**-1
-# Ms_1 trims
-Ms_3 = [np.array([[[-1.19628962e-01,  4.78500000e+00,  1.36125000e+00],       
-                        [-1.22216380e+00, -5.88038994e-17, -1.29732624e-16]],       
-
-                    [[ 2.75448343e+01, -4.09625873e-15, -5.24707922e-16],        
-                        [ 1.38346959e+01, -1.64018392e-15, -9.14780575e-16]],       
-
-                    [[-2.34899407e+00, -5.34145567e-16, -3.02474728e-16],        
-                        [-1.00260753e-15,  1.73733322e-16,  3.97924039e-17]],       
-
-                    [[ 4.43170528e+00,  1.12302734e-16,  4.10866620e-16],        
-                        [ 9.31305763e-16, -1.86261153e-16, -1.86261153e-16]]]), 
-
-            np.array([[[ 2.84845613e+02,  1.65000000e+00],
-                        [ 2.27137940e+02, -1.72605562e-13],
-                        [ 1.29793108e+02,  2.54414728e-14],
-                        [ 4.86724156e+01,  6.09220633e-15],
-                        [ 1.08160924e+01,  1.01171420e-14],
-                        [ 1.08160924e+00, -3.13565083e-15]],
-
-                    [[-9.64091464e+01, -2.86049313e-14],
-                        [ 4.14691757e-14,  7.88666604e-14],
-                        [ 2.67503292e-14, -9.27142549e-15],
-                        [-6.34389840e-14, -6.91944204e-15],
-                        [-1.24333103e-14, -2.02584903e-15],
-                        [ 6.92587964e-15,  8.01648088e-16]],
-
-                    [[ 1.93541862e+02, -5.17408223e-14],
-                        [-5.39535253e-14, -1.96574197e-14],
-                        [-5.00488782e-14, -4.71474082e-14],
-                        [ 2.06908008e-14,  8.68344641e-15],
-                        [ 1.40966336e-14,  2.75871201e-14],
-                        [-3.57269884e-15, -6.35581641e-15]],
-
-                    [[-3.26656222e+01,  1.37816904e-14],
-                        [-1.57824001e-14, -1.92527602e-14],
-                        [ 1.07754349e-15,  7.17861036e-15],
-                        [ 2.10137810e-14, -6.69939041e-16],
-                        [-4.24524051e-15, -2.25717150e-16],
-                        [-3.59823275e-16,  3.49127175e-17]],
-
-                    [[ 4.62211928e+01,  4.21235493e-15],
-                        [-7.00784374e-16,  8.21968518e-15],
-                        [ 4.49894444e-15,  1.37359712e-14],
-                        [ 1.55416701e-15, -5.19175671e-15],
-                        [-2.74045293e-15, -6.88588808e-15],
-                        [ 4.70817493e-16,  1.72633081e-15]]]), 
-
-            np.array([[[ 4.68071678e+00, -4.78500000e+00, -1.36125000e+00],
-                        [ 6.25613546e+00, -2.64267077e-17, -1.19440696e-16],        
-                        [ 1.56403386e+00,  7.59312342e-18,  4.78427451e-17]],       
-
-                    [[-1.20118630e+00, -1.72239045e-16, -1.51455377e-16],        
-                        [ 5.52210332e-16, -3.13691772e-16, -1.22891472e-16],        
-                        [-8.91891458e-17,  9.01323150e-17,  3.09872741e-17]],       
-
-                    [[ 3.39930849e+00,  0.00000000e+00,  3.51928144e-17],        
-                        [ 1.00761893e-16,  0.00000000e+00,  1.00761893e-16],        
-                        [-4.17799856e-17,  0.00000000e+00, -4.17799856e-17]]])]
-errors_3 = np.array([132000.,5000.,2.])
-# 1d trim
-Ms_4 = [np.array([3.432123,4.123,.000000000000001234,.00000000012343421,.0623042345])]
-errors_4 = np.array([24500.])
-# 1d doesn't trim
-Ms_5 = [np.array([3.432123,4.123,.000000000000001234,.00000000012343421,.0623042345])]
-errors_5 = np.array([24.])
-# trimmable, but shoudn't trim because degree is too small to trim so it doesn't trim
-Ms_6 = [1.0*np.arange(3*3*3*3,0,-1).reshape(3,3,3,3),1.0*np.arange(3*3*3*3,0,-1).reshape(3,3,3,3),1.0*np.arange(3*3*3*3,0,-1).reshape(3,3,3,3)]
-errors_6 = np.array([450000000.,450043230000.,45000000000.])
-
-Ms_7 = [1.0*np.arange(3*5*5*5,0,-1).reshape(3,5,5,5),1.0*np.arange(5*5*4*5,0,-1).reshape(5,5,4,5),1.0*np.ones(5*5*3*5).reshape(5,5,3,5),1.0*np.ones(3*3*3*3).reshape(3,3,3,3)]
-errors_7 = np.array([19000000.,19000000.,19000000.,1900.])
-#Obvious which dimension should trim first
-Ms_8 = [np.array([[1,1,1,1,5e-3],[1,1,1,1,5e-3],[1,1,1,1,5e-3],[1,1,1,1,5e-3],[6e-3,5e-3,5e-3,5e-3,5e-3]]),np.array([[1,1,1,1,5e-3],[1,1,1,1,5e-3],[1,1,1,1,5e-3],[1,1,1,1,5e-3],[0,0,0,0,25e-3]]),np.array([[1,1,1,0,0],[1,1,1,0,0],[1,1,1,0,13e-3],[0,0,0,0,0],[13e-3,0,0,0,0]]),np.array([[1,1,1,1,5e-3],[1,1,1,1,5e-3],[1,1,1,1,5e-3],[1,1,1,1,5e-3],[5e-3,5e-3,5e-3,5e-3,5e-3]]),np.array([[0,0,0,0,0],[0,13e-3,0,0,0],[0,0,1e-3,0,0],[0,0,0,13e-3,0],[0,0,0,0,0]])]
-errors_8 = np.array([0.,0.,0.,0.,0.])
-
-ALL_MS = [Ms_4, Ms_3, Ms_6, Ms_7]
-ALL_ERRORS = [errors_4, errors_3, errors_6, errors_7]
